refactor(router): разнос service.go по темам — чистые перемещения
Этап B реструктуризации backend: internal/singbox/router/service.go
(3220 строк, 90+ методов) разнесён дословно (AST-байтовые диапазоны,
включая doc-комментарии) по темам, в конвенции уже существовавших
service_dns.go / service_fakeip_crud.go / service_selective.go:

- service_lifecycle.go — Enable/Disable/Reconcile/GetStatus, netfilter,
  tproxy-inbound heal, ожидание sing-box, emit-события
- service_rules.go — CRUD правил маршрутизации и rule-set'ов
- service_outbounds.go — CRUD композитных outbound'ов, переименование
  тегов, поиск ссылок
- service_settings.go — Get/UpdateSettings, нормализация и валидация
  (fakeip/QoS/ingress), списки WAN/bindable-интерфейсов, пресеты
- service_issues.go — вычисление issues, политики, Inspect/InspectDNS
- service_staging.go — staging-пайплайн (Apply/Discard/Status)
- service.go (ядро, 662 строки) — типы/интерфейсы Deps, конструктор,
  загрузка/persist конфига

Эквивалентность верифицирована мультимножествами нормализованных строк
кода и комментариев. Поведение не меняется.
