+++ internal/singbox/router/service.go
@@ -4,13 +4,9 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"net/netip"
 	"os"
 	"path/filepath"
-	"slices"
-	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -21,9 +17,7 @@ import (
 	"github.com/hoaxisr/awg-manager/internal/presets"
 	"github.com/hoaxisr/awg-manager/internal/singbox/heavyop"
 	"github.com/hoaxisr/awg-manager/internal/singbox/orchestrator"
-	"github.com/hoaxisr/awg-manager/internal/singbox/router/selective"
 	"github.com/hoaxisr/awg-manager/internal/storage"
-	"github.com/hoaxisr/awg-manager/internal/tunnel"
 )
 
 type Service interface {
@@ -650,305 +644,6 @@ func (s *ServiceImpl) orchestratorApplyNow() error {
 	return s.deps.Orch.ReloadNow()
 }
 
-// xtDscpUsable reports whether iptables `-m dscp` matching is usable,
-// honouring the test seam (Deps.XtDscpProbe) when set. Availability changes
-// are logged on TRANSITIONS only (available↔unavailable), never per call —
-// the reconcile loop hits this every tick and a missing optional module must
-// not spam two Warn lines per tick forever (negative probe results are
-// additionally TTL-cached inside IsXtDscpAvailable).
-func (s *ServiceImpl) xtDscpUsable(ctx context.Context) bool {
-	ok := false
-	if s.deps.XtDscpProbe != nil {
-		ok = s.deps.XtDscpProbe(ctx)
-	} else {
-		ok = IsXtDscpAvailable(ctx)
-	}
-	state := int32(2)
-	if ok {
-		state = 1
-	}
-	if prev := s.xtDscpState.Swap(state); prev != state {
-		switch {
-		case !ok:
-			s.appLog.Warn("qos-dscp", "", "xt_dscp недоступен — классы QoS пропущены (см. статус xtDscpAvailable)")
-		case prev == 2:
-			s.appLog.Info("qos-dscp", "", "xt_dscp снова доступен — классы QoS будут применены")
-		}
-	}
-	return ok
-}
-
-// prepareNetfilter runs the common netfilter preflight: xt_TPROXY module
-// load and TPROXY target availability check. It is shared by Enable and
-// reconcileInstalled so both paths run identical validation. Tests can
-// override it via deps.NetfilterPreflight to avoid real syscalls.
-func (s *ServiceImpl) prepareNetfilter(ctx context.Context) error {
-	if s.deps.NetfilterPreflight != nil {
-		return s.deps.NetfilterPreflight(ctx)
-	}
-
-	if err := EnsureTProxyModule(ctx); err != nil {
-		return err
-	}
-
-	if !IsTProxyTargetAvailable(ctx) {
-		return fmt.Errorf("iptables TPROXY target unavailable — kernel module loaded but iptables extension missing")
-	}
-
-	// Best-effort preload of all remaining router netfilter modules.
-	// TPROXY is already handled above as fatal; the rest are soft.
-	// This mirrors the full matcher/target set bisect-combo.sh warms up:
-	// xt_comment, xt_mark, xt_connmark, xt_conntrack, xt_pkttype.
-	if errs := EnsureRouterNetfilterModules(ctx); len(errs) > 0 {
-		for _, err := range errs {
-			s.appLog.Warn("ensure-netfilter", "", err.Error())
-		}
-	}
-
-	return nil
-}
-
-// waitForSingbox polls until sing-box is BOTH process-alive and actually
-// listening on the router inbound sockets (TCP RedirectPort + UDP
-// TPROXYPort), or the deadline expires. Used by Enable after SetEnabled
-// triggers the orchestrator's debounced cold-start so iptables redirects
-// don't land on a TPROXY port that nothing is listening on yet.
-//
-// PID-alive alone is not enough (issue #354): the router config reaches
-// sing-box via the orchestrator's debounced (250ms) reload, so an
-// already-running process keeps serving the OLD inbound set for a moment,
-// and a freshly started one binds inbounds only at the end of startup
-// (after config parse + rule-set load — seconds on mipsel). Gating on the
-// same socket probe GetStatus uses means the status emitted at the end of
-// Enable reflects a truly active interception path instead of a transient
-// «СБОЙ».
-//
-// Returns ctx.Err on cancellation, or a timeout error after the deadline;
-// callers can treat the timeout as soft (proceed with iptables and accept
-// the brief race) or hard at their discretion.
-// fakeIPIfaceName builds the KERNEL interface name for a fakeip-tun OpkgTun from
-// its allocated index (e.g. index 3 → "opkgtun3"). Use this ONLY where the
-// kernel sees the iface: the sing-box tun inbound interface_name, the
-// "ip addr flush dev <iface>" exec, /sys/class/net/<iface>/carrier, the
-// /proc/net/route iface match, and the /sys index scan. For NDMS RCI calls use
-// fakeIPNDMSName instead — NDMS rejects the lowercase kernel name.
-func fakeIPIfaceName(index int) string {
-	return tunnel.NewNames("awg" + strconv.Itoa(index)).IfaceName
-}
-
-// fakeIPNDMSName builds the NDMS RCI interface name for a fakeip-tun OpkgTun from
-// its allocated index (e.g. index 3 → "OpkgTun3"). This mirrors
-// tunnel.Names.NDMSName (CamelCase "OpkgTun%s"); the kernel name is its lowercase
-// (strings.ToLower → fakeIPIfaceName). NDMS REQUIRES this CamelCase form for every
-// RCI interface op (create/delete, address/mtu, up/down) and StaticRouteSpec
-// Interface — passing the lowercase kernel name yields
-// "unsupported interface type: \"opkgtun\"" (stand-verified). Use fakeIPIfaceName
-// only for the kernel-facing sites (sing-box config, ip exec, /sys, /proc).
-func fakeIPNDMSName(index int) string {
-	return tunnel.NewNames("awg" + strconv.Itoa(index)).NDMSName
-}
-
-// ReapOrphanedFakeIPTun removes a fakeip-tun OpkgTun left provisioned by a crash
-// or incomplete teardown when the router is no longer in fakeip-tun mode. It is
-// STARTUP-ONLY (wired once in cmd/awg-manager) — it must NOT run on every
-// Reconcile, or a live mode-switch (fakeip→tproxy) would tear the iface down
-// bluntly, bypassing the safe drain in Disable(fakeip-tun). Idempotent and
-// best-effort: reaps strictly by persisted state (Index). A description-based
-// scan to catch OpkgTuns whose persist was lost is a deferred fallback (v1
-// reaps by persist only).
-//
-// It ALSO sweeps a stale v4 drain reject route for the configured pool in
-// non-fakeip mode — the startup safety net for a disable drain interrupted by a
-// restart (the async drain goroutine does not survive one) or an async-remove
-// that didn't match the route (Fix 1). That sweep is persist-independent.
-//
-// INVARIANT (relied on by this reap): Enable(fakeip-tun) MUST persist the index
-// via SetFakeIPState BEFORE CreateOpkgTun (and roll back its own partial work on
-// failure), so persisted state is a reliable superset of live ifaces — otherwise
-// a crash mid-Enable leaves a persist-less orphan this reap cannot see (the
-// description-scan fallback is deferred).
-func (s *ServiceImpl) ReapOrphanedFakeIPTun(ctx context.Context) error {
-	settings, err := s.deps.Settings.Load()
-	if err != nil {
-		return err
-	}
-	sr, _ := NormalizeSingboxRouterSettings(settings.SingboxRouter)
-	if sr.RoutingMode == "fakeip-tun" {
-		return nil // active mode owns the iface; Enable/Reconcile manage it
-	}
-
-	// Startup safety net for the disable drain (Fix 1): the async drain goroutine
-	// that removes the v4 reject route does NOT survive a daemon restart (no
-	// persisted pending-drain). So in NON-fakeip mode (we are not currently the
-	// active owner) best-effort remove a stale drain reject route for the
-	// CONFIGURED pool. This covers a drain that was interrupted by a restart OR an
-	// async-remove that didn't match the route. Derive net/mask exactly as
-	// disableFakeIPTun does (Masked → splitCIDR). NDMS no:true on a non-existent
-	// route is idempotent, so this is safe on a clean boot. Done independently of
-	// the persist-based iface reap below (the reject route can outlive the FakeIP
-	// persist that disableFakeIPTun clears).
-	//
-	// The reject route is a kill-switch FLAG on the pool→OpkgTun route (stand-
-	// verified), so its NDMS form is interface-bound. We can only target it when the
-	// persisted index tells us the OpkgTun NDMS name; with no persisted index the
-	// route cannot be addressed here (it is tied to the OpkgTun being reaped below,
-	// so DeleteOpkgTun may cascade-remove it — verify cascade semantics on stand).
-	if s.deps.StaticRoutes != nil {
-		if poolNet, poolMask, derr := poolV4NetMask(s.deps.FakeIPTun.Inet4Range); derr == nil {
-			sweepIface := ""
-			if st := settings.FakeIP; st != nil && st.Provisioned {
-				sweepIface = fakeIPNDMSName(st.Index)
-			}
-			if sweepIface != "" {
-				if err := s.deps.StaticRoutes.RemoveStaticRoute(ctx, StaticRouteSpec{
-					Network: poolNet, Mask: poolMask, Interface: sweepIface, Comment: fakeIPDrainComment,
-				}); err != nil {
-					s.appLog.Warn("fakeip-reap", sweepIface, "sweep stale drain reject route: "+err.Error())
-				}
-			}
-		}
-	}
-
-	st := settings.FakeIP
-	if st == nil || !st.Provisioned {
-		return nil // nothing persisted to reap
-	}
-	ndmsName := fakeIPNDMSName(st.Index)
-	if s.deps.OpkgTun == nil {
-		// No provisioner (degraded/test): we can't confirm the iface is gone, so
-		// KEEP the persist — clearing it would convert a tracked orphan into an
-		// un-reapable persist-less one. The index isn't leaked: the allocator is
-		// live-sourced (reads /sys + NDMS), so a still-live iface stays occupied.
-		// A future boot with a real provisioner reaps it.
-		return nil
-	}
-	if err := s.deps.OpkgTun.DeleteOpkgTun(ctx, ndmsName); err != nil {
-		// Keep the persist on failure: returning without clearing it lets the
-		// next boot retry the reap rather than leaking the orphan forever.
-		return fmt.Errorf("reap opkgtun %s: %w", ndmsName, err)
-	}
-	s.appLog.Info("fakeip-reap", ndmsName, "removed orphaned fakeip OpkgTun (mode != fakeip-tun)")
-	// Clear persist ONLY after a confirmed delete success (NDMS returns nil even
-	// when the iface was already gone, i.e. idempotent), so the index frees.
-	return s.deps.Settings.SetFakeIPState(nil)
-}
-
-// fakeIPReadyInputs derives the inputs the fakeip-tun readiness probes need
-// from loaded settings + the static FakeIPTun params: the tun iface name (from
-// the allocated OpkgTun index), the tun-side DNS address (the other /30 host,
-// where sing-box's DNS server listens), and the fakeip v4 pool prefix. ok is
-// false when fakeip is not provisioned or any field is unparseable, so callers
-// can fail-closed without a fakeip branch firing on tproxy state.
-func (s *ServiceImpl) fakeIPReadyInputs() (iface, dnsAddr string, fakeipNet netip.Prefix, ok bool) {
-	if s.deps.Settings == nil {
-		return "", "", netip.Prefix{}, false
-	}
-	settings, err := s.deps.Settings.Load()
-	if err != nil || settings == nil || settings.FakeIP == nil || !settings.FakeIP.Provisioned {
-		return "", "", netip.Prefix{}, false
-	}
-	iface = fakeIPIfaceName(settings.FakeIP.Index)
-	dnsAddr, err = DeriveTunDNS(s.deps.FakeIPTun.TunAddr4)
-	if err != nil {
-		return "", "", netip.Prefix{}, false
-	}
-	fakeipNet, err = netip.ParsePrefix(s.deps.FakeIPTun.Inet4Range)
-	if err != nil || !fakeipNet.Addr().Is4() {
-		return "", "", netip.Prefix{}, false
-	}
-	return iface, dnsAddr, fakeipNet, true
-}
-
-func (s *ServiceImpl) waitForSingbox(ctx context.Context, timeout time.Duration) error {
-	if s.deps.Singbox == nil {
-		return nil
-	}
-
-	// Mode-aware readiness: read the mode INTERNALLY (the signature has test
-	// callers and must not change). fakeip-tun has no inbound sockets, so the
-	// tproxy socket probe never turns true for it — gate instead on process +
-	// tun carrier (carrier=1 = sing-box attached the gvisor tun stack, the
-	// structural "config is live" signal). The live .2→fakeip DNS answer is NO
-	// longer in this gate (it tripped on resolv.conf attempts:1, stand-verified
-	// 2026-06-15) — it is now a best-effort confirm after readiness in
-	// enableFakeIPTun. See singboxReady for the full rationale.
-	fakeIP := false
-	if s.deps.Settings != nil {
-		if settings, err := s.deps.Settings.Load(); err == nil && settings != nil {
-			fakeIP = settings.SingboxRouter.RoutingMode == "fakeip-tun"
-		}
-	}
-
-	deadline := time.Now().Add(timeout)
-	start := time.Now()
-	lastHeartbeat := time.Time{}
-	const pollInterval = 100 * time.Millisecond
-	for {
-		if s.singboxReady(ctx, fakeIP) {
-			return nil
-		}
-		if fn := s.transitionReadinessProgress; fn != nil && time.Since(lastHeartbeat) >= 2*time.Second {
-			elapsed := time.Since(start).Round(time.Second)
-			running, _ := s.deps.Singbox.IsRunning()
-			msg := fmt.Sprintf("запуск sing-box… %s", elapsed)
-			if running {
-				msg = fmt.Sprintf("sing-box работает, ожидаем inbounds… %s", elapsed)
-			}
-			fn(msg)
-			lastHeartbeat = time.Now()
-		}
-		if time.Now().After(deadline) {
-			return fmt.Errorf("sing-box did not come up within %s", timeout)
-		}
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-time.After(pollInterval):
-		}
-	}
-}
-
-// singboxReady reports whether sing-box is up for the active mode. tproxy:
-// process + both inbound sockets bound. fakeip-tun: process + tun carrier.
-//
-// For fakeip-tun, carrier=1 IS the structural readiness signal: it means
-// sing-box created and attached the gvisor tun stack from the fakeip config —
-// the analog of "inbound socket bound" for tproxy, and it is fast and reliable.
-//
-// The live .2→fakeip DNS probe was DEMOTED out of this hard gate (stand-verified
-// 2026-06-15): the Go resolver (net.Resolver{PreferGo:true}) HONORS the router's
-// /etc/resolv.conf `options timeout:1 attempts:1`, so it does a single ~1s-bounded
-// attempt with no retry. In the first seconds after sing-box starts the fakeip
-// round-trip to .2 is occasionally slower than that, so the probe returned false
-// on every poll and waitForSingbox timed out at 60s — falsely failing Enable even
-// though sing-box was fully up (carrier=1) and fakeip worked. The DNS check now
-// runs ONCE as a best-effort, logged confirmation AFTER readiness (see
-// enableFakeIPTun), never as a flaky gate. ctx is unused now that the live DNS
-// probe is out of the gate; kept on the signature for the tproxy/test callers.
-func (s *ServiceImpl) singboxReady(_ context.Context, fakeIP bool) bool {
-	running, _ := s.deps.Singbox.IsRunning()
-	if !running {
-		return false
-	}
-	if !fakeIP {
-		// HARD gate (issue #221): only the procfs socket probe proves the
-		// router-slot TPROXY/REDIRECT inbounds actually bound. A healthy
-		// Clash API is NOT equivalent — the process can be up and serving
-		// Clash while the router inbounds failed to bind (port taken,
-		// rejected hot-reload), and installing iptables in that state
-		// blackholes all policy traffic including DNS:53.
-		return singboxListeningProbe()
-	}
-	// Only iface is needed for the carrier gate; dnsAddr/fakeipNet (which the
-	// demoted DNS probe used) are derived later in enableFakeIPTun for the
-	// best-effort confirm.
-	iface, _, _, ok := s.fakeIPReadyInputs()
-	if !ok {
-		return false
-	}
-	return tunReadyProbe(iface)
-}
-
 func (s *ServiceImpl) persistConfig(ctx context.Context, cfg *RouterConfig) error {
 	materialized, err := s.ruleSetMaterializer().materializeConfig(cfg)
 	if err != nil {
@@ -1038,1119 +733,6 @@ func (s *ServiceImpl) persistConfig(ctx context.Context, cfg *RouterConfig) erro
 	return nil
 }
 
-func cleanValidateError(err error) string {
-	msg := err.Error()
-	msg = strings.ReplaceAll(msg, "\x1b[31m", "")
-	msg = strings.ReplaceAll(msg, "\x1b[0m", "")
-	if idx := strings.Index(msg, "FATAL"); idx >= 0 {
-		msg = msg[idx+len("FATAL"):]
-	}
-	msg = strings.TrimSpace(msg)
-	if idx := strings.Index(msg, ": exit status"); idx > 0 {
-		msg = msg[:idx]
-	}
-	if idx := strings.Index(msg, "decode config at "); idx >= 0 {
-		tail := msg[idx+len("decode config at "):]
-		if j := strings.Index(tail, ": "); j > 0 {
-			tail = tail[j+2:]
-		}
-		msg = "конфиг недопустим: " + tail
-	}
-	return strings.TrimSpace(msg)
-}
-
-// Enable is the USER-INITIATED router enable (HTTP handler + SwitchRoutingMode).
-// It clears the sticky master-stop intent — an explicit enable is an explicit
-// intent to run sing-box, which must override a prior master-Stop — then runs
-// the idempotent provisioning. Drift-heal (Reconcile / reconcileFakeIPTun) must
-// NOT clear the intent (the watchdog must respect a user's manual stop and not
-// resurrect the daemon on drift), so it calls enableLocked(ctx, false) instead.
-func (s *ServiceImpl) Enable(ctx context.Context) error {
-	return s.enableLocked(ctx, true)
-}
-
-// enableLocked provisions the router under s.mu. clearManualStop gates the
-// sticky-intent clear: true for user-initiated Enable, false for drift-heal
-// reuse (Reconcile / reconcileFakeIPTun) which must honour a prior master-Stop.
-func (s *ServiceImpl) enableLocked(ctx context.Context, clearManualStop bool) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	// Validate settings first — fail fast with a meaningful error before
-	// attempting any kernel / iptables operations.
-	settings, err := s.deps.Settings.Load()
-	if err != nil {
-		return err
-	}
-	sr, err := NormalizeSingboxRouterSettings(settings.SingboxRouter)
-	if err != nil {
-		return fmt.Errorf("router settings: %w", err)
-	}
-
-	// Explicit enable = explicit intent to run sing-box. Clear any sticky
-	// manual-stop so the orchestrator cold-start (triggered by SetEnabled below)
-	// isn't suppressed by shouldRun()=!IsManuallyStopped — otherwise Enable waits
-	// the full boot window and fails with a misleading readiness timeout
-	// (stand-found 2026-06-15, applies to BOTH tproxy and fakeip-tun). Gated to
-	// user-initiated Enable: a drift-heal reconcile (clearManualStop=false) must
-	// NOT wipe a user's master-Stop. The nil guard keeps test wirings that omit
-	// Singbox working.
-	if clearManualStop && s.deps.Singbox != nil {
-		if err := s.deps.Singbox.ClearManualStop(); err != nil {
-			return fmt.Errorf("clear manual-stop intent: %w", err)
-		}
-	}
-
-	// fakeip-tun has an entirely separate provisioning path (OpkgTun + tun +
-	// fakeip DNS + pool/CIDR routes) with its own rollback. The tproxy body
-	// below stays byte-for-byte unchanged for RoutingMode=="tproxy".
-	if sr.RoutingMode == "fakeip-tun" {
-		sr.Enabled = true
-		return s.enableFakeIPTun(ctx, settings, sr)
-	}
-
-	policyMode := sr.DeviceMode == "" || sr.DeviceMode == "policy"
-	mark := ""
-	if policyMode {
-		if sr.PolicyName == "" {
-			return ErrPolicyNotConfigured
-		}
-		mark, err = s.deps.Policies.GetPolicyMark(ctx, sr.PolicyName)
-		if err != nil || mark == "" {
-			return fmt.Errorf("policy %q: %w", sr.PolicyName, ErrPolicyMissing)
-		}
-	}
-
-	if err := s.prepareNetfilter(ctx); err != nil {
-		return err
-	}
-
-	sr.Enabled = true
-
-	cfg, err := s.loadRouterConfig()
-	if err != nil {
-		return err
-	}
-	cfg.Inbounds = ensureTProxyInbound(cfg.Inbounds, sr.UDPTimeout)
-	cfg.Outbounds = stripAutoManagedDirect(cfg.Outbounds)
-	cfg.EnsureSystemRules(sr.SnifferEnabled)
-	// Neutralize sing-box's short per-protocol UDP timeouts (QUIC/DTLS 30s,
-	// STUN/DNS 10s) applied on sniff/port inference — they ignore the inbound
-	// udp_timeout and drop games/VoIP early. Raise them to the effective inbound
-	// value via a route-options rule. Placed after the system prefix, before user
-	// rules, so it runs ahead of any final `route` action.
-	cfg.EnsureUDPTimeoutRule(resolveUDPTimeout(sr.UDPTimeout))
-	// QoS-by-DSCP (issue #371): per-class inbound pairs, derived from the
-	// same settings snapshot the iptables spec below uses so ports/classes
-	// cannot drift between the two. The managed route rules live in their
-	// own slot (18-qos-routes.json) and are synced after the config write
-	// below — see qos_routes.go for why they must not live in 20-router.json.
-	qosClasses := activeQoSClasses(sr.QoSClasses)
-	cfg.Inbounds, _ = ensureQoSInbounds(cfg.Inbounds, qosClasses, sr.UDPTimeout)
-	// Settings was already loaded above; revalidate here in case the
-	// store is corrupted or hand-edited around a schema migration. We
-	// fail Enable rather than apply a half-broken config — the user
-	// sees a clean error in the UI and can fix it.
-	if err := ValidateSingboxRouterSettings(sr); err != nil {
-		return fmt.Errorf("router settings: %w", err)
-	}
-	cfg.EnsureRouteWAN(sr.WANAutoDetect, sr.WANInterface)
-
-	// Promote SlotRouter to active FIRST so persistConfigDirect's
-	// orch.Save targets the active path (it keys on the slot's enabled
-	// flag). SetEnabled also triggers the orchestrator's debounced cold-
-	// start — sing-box will read the active config we are about to write.
-	// Legacy fallback (tests) keeps the explicit Start call.
-	if s.deps.Orch != nil {
-		if err := s.deps.Orch.SetEnabled(orchestrator.SlotRouter, true); err != nil {
-			return fmt.Errorf("orchestrator enable router: %w", err)
-		}
-	} else {
-		if running, _ := s.deps.Singbox.IsRunning(); !running {
-			if err := s.deps.Singbox.Start(); err != nil {
-				return fmt.Errorf("sing-box start: %w", err)
-			}
-		}
-	}
-	// Direct write — no staging. Byte-equal short-circuit makes boot
-	// recovery (Reconcile→Enable with iptables gone but active config
-	// already on disk) a no-op write, which is what kills the phantom
-	// "Несохранённые правки" banner that used to follow every reboot.
-	if err := s.persistConfigDirect(ctx, cfg); err != nil {
-		return err
-	}
-	// Managed QoS route rules → 18-qos-routes.json. Runs AFTER the router
-	// slot write so outbound resolution sees the applied config; the
-	// orchestratorApplyNow below covers both slots in one reload.
-	if _, err := s.syncQoSRoutesSlot(ctx, qosClasses); err != nil {
-		return fmt.Errorf("sync qos routes slot: %w", err)
-	}
-	// Слот 20 снова активен — зависимые продюсеры (device-proxy) должны
-	// перегенерировать свои слоты (вернуть ссылки на композиты) ДО reload.
-	s.notifyRoutingSlotsChanged()
-	if err := s.orchestratorApplyNow(); err != nil {
-		return fmt.Errorf("orchestrator reload after enable: %w", err)
-	}
-
-	// Wait for sing-box to be listening before iptables start redirecting
-	// traffic to its TPROXY/REDIRECT ports. HARD fail (issue #221): if
-	// sing-box never comes up — most commonly because a slot config is
-	// rejected by `sing-box check` at load time — installing the AWGM-TPROXY
-	// rule still redirects DNS:53 to 127.0.0.1:<proxy_port>, where nothing
-	// is listening, and the router loses DNS until the user manually stops
-	// awg-manager. The earlier "brief packet-drop blip vs no routing"
-	// trade-off is wrong: a failed sing-box start turns the blip into a
-	// permanent outage.
-	//
-	// Same env-var contract as singbox.maxSingboxBootWait — clamped to a
-	// 60s floor (bootWaitWithFloor). Import-cycle (integration_test in parent
-	// already pulls router) blocks reusing the parent helper directly.
-	bootWait := bootWaitWithFloor()
-	if err := s.waitForSingbox(ctx, bootWait); err != nil {
-		return fmt.Errorf("%w: waited %s (%v)", ErrSingboxNotReady, bootWait, err)
-	}
-
-	// Collect WAN IPs BEFORE Install: the router's own public-IP
-	// addresses on default-route interfaces become RETURN rules in
-	// AWGM-TPROXY/AWGM-REDIRECT, preventing LAN-to-router-WAN-IP
-	// traffic from looping back into sing-box. A collector failure
-	// is fatal — installing without the exclusions would silently
-	// expose the loop edge case to users.
-	wanIPs, err := s.deps.WANIPCollector.Collect(ctx)
-	if err != nil {
-		return fmt.Errorf("collect WAN IPs: %w", err)
-	}
-
-	// Discover LAN bridges that NDMS knows how to REDIRECT DNS for
-	// (i.e. has _NDM_HOTSPOT_DNSREDIR rules on). DNS-NOPOLICY rules
-	// re-mark mark=0 DNS up to one of those marks so the existing
-	// REDIRECT picks it up and forwards to the per-policy ndnproxy.
-	// We pass our policy mark so the picker avoids it — re-marking
-	// default DNS up to the sing-box mark would route it via Policy1's
-	// (permit-less) table and DNS would never resolve. Empty result =
-	// no qualifying bridges = skip the DNS-NOPOLICY logic entirely.
-	var lanBridges []LANBridgeDNSRedir
-	if policyMode {
-		lanBridges, _ = DiscoverLANBridges(ctx, mark)
-		if len(lanBridges) == 0 {
-			s.appLog.Warn("discover-lan-bridges", "", "no NDMS hotspot LAN bridges, DNS fallback skipped")
-		}
-	}
-
-	var ingress []string
-	if policyMode {
-		ingress = s.resolveIngressInterfaces(ctx, sr.IngressInterfaces)
-	}
-
-	bypassUDP, bypassTCP, _ := resolveBypassPorts(sr.BypassPresets, sr.BypassExtraPorts)
-	bypassSubnets, _ := resolveBypassCIDRs(sr.BypassPresets, sr.BypassExtraSubnets)
-
-	// Pre-create the AWGM-SELECTIVE ipset (empty) before iptables-restore
-	// when selective bypass is enabled. iptables-restore fails immediately
-	// if the set referenced in -m set --match-set doesn't exist yet.
-	if sr.SelectiveBypass {
-		if err := ensureSelectiveSetExists(ctx); err != nil {
-			s.appLog.Warn("selective", "", fmt.Sprintf("pre-create ipset failed: %v", err))
-		}
-	}
-
-	// QoS iptables dispatch — graceful degradation: when xt_dscp support is
-	// missing the DSCP rules are skipped (feature-off) with a warning, NEVER
-	// failing Enable — otherwise a missing optional module would take down
-	// the whole interception path at iptables-restore COMMIT (XKeen shipped
-	// the same policy). EnsureXtDscpModule first: prepareNetfilter already
-	// tried, this is the belt-and-suspenders retry closest to Install.
-	qosSpecs := qosIPTablesSpecs(qosClasses)
-	if len(qosSpecs) > 0 {
-		if err := EnsureXtDscpModule(ctx); err != nil {
-			s.appLog.Warn("ensure-xt-dscp", "", err.Error())
-		}
-		if !s.xtDscpUsable(ctx) {
-			s.appLog.Warn("qos-dscp", "", "xt_dscp недоступен — классы QoS пропущены (см. статус xtDscpAvailable)")
-			qosSpecs = nil
-		}
-	}
-
-	if err := s.deps.IPTables.Install(ctx, RestoreInputSpec{
-		PolicyMark:        mark,
-		MatchAll:          !policyMode,
-		WANIPs:            wanIPs,
-		LANBridges:        lanBridges,
-		BypassUDPPorts:    bypassUDP,
-		BypassTCPPorts:    bypassTCP,
-		BypassCIDRs:       bypassSubnets,
-		IngressInterfaces: ingress,
-		SelectiveIPSet:    sr.SelectiveBypass,
-		QoSClasses:        qosSpecs,
-	}); err != nil {
-		// Stop sing-box from listening on the now-orphan TPROXY port,
-		// but DO NOT corrupt the persisted user config. With orchestrator
-		// wired we just park the slot back under disabled/ — sing-box
-		// stops seeing it on next reload, the file's content (including
-		// tproxy-in) is preserved verbatim. Without the orchestrator
-		// (legacy fallback) the only recourse is to strip the inbound.
-		if s.deps.Orch != nil {
-			_ = s.deps.Orch.SetEnabled(orchestrator.SlotRouter, false)
-			// The QoS overlay references qos-* inbounds that just got
-			// parked with the router slot — park it too.
-			_ = s.disableQoSRoutesSlot()
-			// Слот 20 снова выключен — device-proxy должен деградировать
-			// ссылки на композиты до дефолт-членов до ближайшего reload.
-			s.notifyRoutingSlotsChanged()
-		} else {
-			cfg.Inbounds = filterTProxyInbound(cfg.Inbounds)
-			_ = s.persistConfigDirect(ctx, cfg)
-		}
-		return fmt.Errorf("iptables install: %w", err)
-	}
-	s.currentMark = mark
-	s.currentWANIPs = wanIPs
-	s.currentLANBridges = lanBridges
-	s.currentBypassPresets = sr.BypassPresets
-	s.currentBypassExtraPorts = sr.BypassExtraPorts
-	s.currentBypassExtraSubnets = sr.BypassExtraSubnets
-	s.currentIngress = ingress
-	s.currentSelectiveBypass = sr.SelectiveBypass
-	s.currentQoSClasses = qosSpecs
-	s.netfilterStateKnown = true
-
-	settings.SingboxRouter = sr
-	if err := s.deps.Settings.Save(settings); err != nil {
-		return err
-	}
-
-	// Populate the freshly-created (empty) AWGM-SELECTIVE set right away.
-	// Without this, everything between Enable and the first reconcile-driven
-	// rebuild (startup delay + tick, minutes) matches nothing in the guard
-	// and "proxied" traffic leaves via WAN in the clear.
-	if sr.SelectiveBypass {
-		s.triggerSelectiveRebuild(ctx)
-	}
-
-	s.emitStatus(ctx)
-	return nil
-}
-
-func filterTProxyInbound(in []Inbound) []Inbound {
-	out := make([]Inbound, 0, len(in))
-	for _, i := range in {
-		if i.Tag != "tproxy-in" {
-			out = append(out, i)
-		}
-	}
-	return out
-}
-
-// healTProxyInbound checks the persisted router config and brings the
-// tproxy-in inbound to spec: re-adds it if missing, and applies the current
-// udpTimeout if it drifted (e.g. the user changed the setting while the engine
-// was running — this is the Reconcile path that path takes). Idempotent.
-func (s *ServiceImpl) healTProxyInbound(ctx context.Context, udpTimeout string) error {
-	cfg, err := s.loadRouterConfig()
-	if err != nil {
-		return err
-	}
-	// Cheap steady-state guard: present and already at the desired timeout →
-	// skip the marshal/write entirely (this runs on every reconcile tick).
-	effective := resolveUDPTimeout(udpTimeout)
-	for _, in := range cfg.Inbounds {
-		if in.Tag == "tproxy-in" {
-			if in.UDPTimeout == effective {
-				return nil
-			}
-			break // present but drifted — fall through to re-apply
-		}
-	}
-	cfg.Inbounds = ensureTProxyInbound(cfg.Inbounds, udpTimeout)
-	// System self-heal — direct write, no staging UI.
-	return s.persistConfigDirect(ctx, cfg)
-}
-
-// ensureTProxyInbound enforces the SKeen-style split: tproxy-in
-// handles UDP only, redirect-in handles TCP. TPROXY for TCP relies on
-// `-m socket --transparent` to deliver established-connection packets
-// to sing-box's accept()ed transparent socket, but that match
-// evaluates to 0 on Keenetic 4.9-ndm-5 — established TCP packets fall
-// through to the listener and get RST. NAT REDIRECT sidesteps the
-// problem: conntrack records the DNAT for SYN, established packets
-// are auto-translated.
-//
-// Both inbounds bind to 0.0.0.0 because iptables REDIRECT rewrites
-// the packet destination to the *primary IP of the inbound interface*
-// (e.g. 10.10.10.1 on br0), NOT to 127.0.0.1. A listener on 127.0.0.1
-// would never see redirected packets — kernel emits RST. SKeen uses
-// "::" for the same reason.
-const inboundListen = "0.0.0.0"
-
-// DefaultUDPTimeout is the fallback UDP session timeout when the user has not
-// configured a custom value. It matches sing-box's built-in C.UDPTimeout (5m):
-// fakeip's tun-in previously carried no udp_timeout and thus ran at the engine's
-// 5m, so defaulting to 5m here keeps unconfigured sessions no shorter than
-// before while still letting the user raise it. Shorter values dropped games /
-// VoIP that go quiet mid-session.
-const DefaultUDPTimeout = "5m0s"
-
-// resolveUDPTimeout returns the effective UDP timeout string: the user value
-// when non-empty, otherwise DefaultUDPTimeout.
-func resolveUDPTimeout(configured string) string {
-	if configured != "" {
-		return configured
-	}
-	return DefaultUDPTimeout
-}
-
-func ensureTProxyInbound(in []Inbound, udpTimeout string) []Inbound {
-	effective := resolveUDPTimeout(udpTimeout)
-	hasTProxy := false
-	hasRedirect := false
-	for i := range in {
-		switch in[i].Tag {
-		case "tproxy-in":
-			hasTProxy = true
-			// Force UDP-only on existing entry. Older configs had no
-			// `network` field which means TCP+UDP — that's the broken
-			// behaviour we're moving away from.
-			if in[i].Network != "udp" {
-				in[i].Network = "udp"
-			}
-			if !in[i].UDPFragment {
-				in[i].UDPFragment = true
-			}
-			// Always apply the effective timeout — user may have changed it.
-			in[i].UDPTimeout = effective
-			// tcp_fast_open is meaningless on a UDP-only inbound.
-			if in[i].TCPFastOpen {
-				in[i].TCPFastOpen = false
-			}
-			// Strip RoutingMark — see history note below.
-			if in[i].RoutingMark != 0 {
-				in[i].RoutingMark = 0
-			}
-			if in[i].Listen != inboundListen {
-				in[i].Listen = inboundListen
-			}
-		case "redirect-in":
-			hasRedirect = true
-			if !in[i].TCPFastOpen {
-				in[i].TCPFastOpen = true
-			}
-			if in[i].Listen != inboundListen {
-				in[i].Listen = inboundListen
-			}
-		}
-	}
-	out := in
-	if !hasTProxy {
-		out = append([]Inbound{{
-			Type:        "tproxy",
-			Tag:         "tproxy-in",
-			Listen:      inboundListen,
-			ListenPort:  TPROXYPort,
-			Network:     "udp",
-			UDPFragment: true,
-			UDPTimeout:  effective,
-		}}, out...)
-	}
-	if !hasRedirect {
-		out = append([]Inbound{{
-			Type:        "redirect",
-			Tag:         "redirect-in",
-			Listen:      inboundListen,
-			ListenPort:  RedirectPort,
-			TCPFastOpen: true,
-		}}, out...)
-	}
-	return out
-}
-
-func (s *ServiceImpl) emitStatus(ctx context.Context) {
-	if s.deps.Events == nil {
-		return
-	}
-	status, _ := s.GetStatus(ctx)
-	s.deps.Events.Publish("singbox-router:status", status)
-}
-
-func (s *ServiceImpl) emitStagingEvent(reason string) {
-	if s.deps.Bus == nil {
-		return
-	}
-	s.deps.Bus.Publish("resource:invalidated", map[string]any{
-		"resource": "singbox.router.staging",
-		"reason":   reason,
-	})
-}
-
-func (s *ServiceImpl) emitRulesEvent() {
-	if s.deps.Bus == nil {
-		return
-	}
-	s.deps.Bus.Publish("resource:invalidated", map[string]any{
-		"resource": "singbox.router.rules",
-	})
-}
-
-func (s *ServiceImpl) GetStatus(ctx context.Context) (Status, error) {
-	settings, _ := s.deps.Settings.Load()
-	sr := storage.SingboxRouterSettings{}
-	if settings != nil {
-		sr, _ = NormalizeSingboxRouterSettings(settings.SingboxRouter)
-	}
-	cfg, _ := s.loadRouterConfigForMode(sr.RoutingMode)
-	if cfg == nil {
-		cfg = NewEmptyConfig()
-	}
-	awgCount := 0
-	compCount := len(cfg.CompositeOutbounds())
-
-	policyExists := false
-	policyMark := ""
-	deviceCount := 0
-	if sr.PolicyName != "" && s.deps.Policies != nil {
-		if mark, err := s.deps.Policies.GetPolicyMark(ctx, sr.PolicyName); err == nil && mark != "" {
-			policyExists = true
-			policyMark = mark
-		}
-		if devices, err := s.deps.Policies.ListDevicesForPolicy(ctx, sr.PolicyName); err == nil {
-			for _, d := range devices {
-				if d.Bound {
-					deviceCount++
-				}
-			}
-		}
-	}
-
-	// One -S probe per table yields both chain existence and jump presence.
-	// A probe error is treated as "unknown" — installed/jumps stay false but
-	// the badge self-corrects on the next status read (no side effect here,
-	// unlike the reconcile path which must not reinstall on a transient error).
-	installed, jumps, _ := s.deps.IPTables.Probe(ctx)
-	// Active = interception path truly live, computed per routing mode.
-	var active bool
-	if sr.RoutingMode == "fakeip-tun" {
-		// fakeip-tun has no iptables jumps and no inbound sockets. Steady-state
-		// liveness = process up + tun carrier + the fakeip pool auto-route
-		// present (the honest structural check — the fakeip equivalent of
-		// "TPROXY jumps present"). No live DNS query here: that would add
-		// per-poll latency, and the route-presence check is enough once Enable
-		// has finished wiring routes (waitForSingbox already gated on live DNS).
-		running, _ := s.deps.Singbox.IsRunning()
-		if iface, _, fakeipNet, ok := s.fakeIPReadyInputs(); ok {
-			active = running && tunReadyProbe(iface) && fakeIPPoolRoutePresent(iface, fakeipNet)
-		}
-	} else {
-		// tproxy: chains + PREROUTING jumps + sing-box listening on both inbound sockets.
-		active = jumps && singboxListeningProbe()
-	}
-	// Surface the captured sing-box fatal reason only when the engine is
-	// meant to be up but isn't (СБОЙ). lastError is cleared by the operator
-	// on a successful (re)start, so a healthy engine reports empty.
-	lastError := ""
-	if sr.Enabled && !active && s.deps.Singbox != nil {
-		lastError = s.deps.Singbox.LastError()
-	}
-	// Crash observability (#456): счётчик недавних падений, причина
-	// последнего и пауза авто-перезапуска. Заполняется всегда (omitempty
-	// прячет нули) — UI показывает блок и после восстановления, пока
-	// падения не выйдут из окна.
-	crashCount := 0
-	lastCrashReason := ""
-	restartSuppressedUntil := ""
-	var suppressedUntil time.Time
-	if s.deps.Singbox != nil {
-		n, reason, until := s.deps.Singbox.CrashStats()
-		crashCount = n
-		lastCrashReason = reason
-		if !until.IsZero() {
-			suppressedUntil = until
-			restartSuppressedUntil = until.Format(time.RFC3339)
-		}
-	}
-	issues := s.computeIssues(cfg)
-	// Мёртвый движок при живом перехвате (#456 FIX-B): PREROUTING-джампы
-	// стоят, а процесс не работает — весь policy-трафик (включая hijacked
-	// DNS:53) уходит в мёртвый порт до конца backoff-паузы. computeIssues
-	// видит только конфиг, поэтому этот runtime-issue собирается здесь, где
-	// уже есть probe и crash-статистика. Только tproxy: у fakeip-tun нет
-	// iptables-перехвата.
-	if sr.Enabled && sr.RoutingMode != "fakeip-tun" && jumps && s.deps.Singbox != nil {
-		if running, _ := s.deps.Singbox.IsRunning(); !running {
-			msg := "Движок остановлен, но перехват трафика активен — трафик политик не ходит."
-			if !suppressedUntil.IsZero() {
-				msg += fmt.Sprintf(" Автоперезапуск приостановлен до %s (падений за 10 мин: %d).",
-					suppressedUntil.Local().Format("15:04"), crashCount)
-			} else {
-				msg += " Автоперезапуск: при следующей проверке (до 30 с)."
-			}
-			issues = append(issues, Issue{
-				Severity: "error",
-				Kind:     "engine-dead-interception",
-				Message:  msg,
-			})
-		}
-	}
-	// Эксперт-редактор (90-user.json): если последний reload пропущен из-за
-	// провала кросс-слот валидации по вине пользовательского слота, движок
-	// продолжает работать на старом конфиге, а сам файл оркестратор
-	// намеренно не чинит (prune пропускает user-слот). computeIssues видит
-	// только конфиг роутера, поэтому runtime-issue собирается здесь из
-	// orchestrator.LastReloadValidation — по паттерну #456.
-	if s.deps.Orch != nil {
-		if v := s.deps.Orch.LastReloadValidation(); v != nil {
-			for _, ve := range v.Errors {
-				if ve.Slot != orchestrator.SlotUser || ve.Severity == orchestrator.SeverityWarning {
-					continue
-				}
-				var msg string
-				switch {
-				case strings.HasPrefix(ve.Kind, "unknown-") && ve.Tag != "":
-					msg = fmt.Sprintf("Пользовательский конфиг (90-user.json) ссылается на несуществующий тег %q — правьте в редакторе конфигурации", ve.Tag)
-				case ve.Tag != "":
-					msg = fmt.Sprintf("Пользовательский конфиг (90-user.json): %s %q — правьте в редакторе конфигурации", ve.Kind, ve.Tag)
-				default:
-					msg = fmt.Sprintf("Пользовательский конфиг (90-user.json): %s — правьте в редакторе конфигурации", ve.Message)
-				}
-				issues = append(issues, Issue{
-					Severity: "error",
-					Kind:     "user-slot-validation",
-					Tag:      ve.Tag,
-					Message:  msg,
-				})
-			}
-		}
-	}
-	// QoS-DSCP support: xtDscpAvailable is always reported (the UI keys the
-	// feature's "supported" state on it). When classes are actually
-	// configured but the support probe fails, additionally surface an issue
-	// that distinguishes the two causes (kernel module vs iptables
-	// extension) so diagnostics can tell them apart. The detailed check on
-	// the failure path is served from the same TTL-bounded probe cache as
-	// the availability flag, so status polling never execs iptables per poll.
-	qosActive := activeQoSClasses(sr.QoSClasses)
-	// A class whose outbound no longer resolves is skipped at emit time
-	// (syncQoSRoutesSlot) — surface WHY the class is inert so the user can
-	// re-point or disable it.
-	if sr.RoutingMode != "fakeip-tun" {
-		for _, c := range qosActive {
-			if !s.isKnownOutboundTag(ctx, c.Outbound, cfg) {
-				issues = append(issues, Issue{
-					Severity: "warning",
-					Kind:     "qos-outbound-missing",
-					Tag:      c.Outbound,
-					Message:  fmt.Sprintf("класс QoS (DSCP %d) ссылается на несуществующий outbound %q — класс не применяется", c.DSCP, c.Outbound),
-				})
-			}
-		}
-	}
-	xtDscpAvailable := s.xtDscpUsable(ctx)
-	if !xtDscpAvailable && sr.RoutingMode != "fakeip-tun" && len(qosActive) > 0 {
-		moduleOK, matchOK := cachedXtDscpAvailability(ctx)
-		var msg string
-		switch {
-		case !moduleOK && !matchOK:
-			msg = "QoS DSCP: модуль ядра xt_dscp не найден и расширение iptables «dscp» недоступно — классы QoS не будут применены"
-		case !moduleOK:
-			msg = "QoS DSCP: модуль ядра xt_dscp не найден (/lib/modules) — классы QoS не будут применены"
-		default:
-			msg = "QoS DSCP: расширение iptables «dscp» недоступно — классы QoS не будут применены"
-		}
-		issues = append(issues, Issue{Severity: "warning", Kind: "qos-xt-dscp", Message: msg})
-	}
-	// fakeip-tun active iface: surface the provisioned kernel iface name
-	// ("opkgtun<idx>") so the UI can show it in the engine-settings panel. Only
-	// when in fakeip-tun mode AND actually provisioned (persisted FakeIPState);
-	// empty otherwise.
-	var fakeIPIface string
-	fakeIPDns := ""
-	fakeIPTunAddr := ""
-	if sr.RoutingMode == "fakeip-tun" && settings != nil &&
-		settings.FakeIP != nil && settings.FakeIP.Provisioned {
-		fakeIPIface = fakeIPIfaceName(settings.FakeIP.Index)
-		if d, derr := DeriveTunDNS(s.deps.FakeIPTun.TunAddr4); derr == nil {
-			fakeIPDns = d
-		}
-		if addr, _, aerr := splitCIDRToAddrMask(s.deps.FakeIPTun.TunAddr4); aerr == nil {
-			fakeIPTunAddr = addr
-		}
-	}
-	return Status{
-		Enabled:                sr.Enabled,
-		Installed:              installed,
-		Active:                 active,
-		NetfilterAvailable:     IsNetfilterAvailable(),
-		NetfilterComponentName: "Модули ядра подсистемы сетевой фильтрации",
-		TProxyTargetAvailable:  IsTProxyTargetAvailable(ctx),
-		XtDscpAvailable:        xtDscpAvailable,
-		PolicyName:             sr.PolicyName,
-		PolicyMark:             policyMark,
-		PolicyExists:           policyExists,
-		DeviceMode:             sr.DeviceMode,
-		SnifferEnabled:         sr.SnifferEnabled,
-		DeviceCount:            deviceCount,
-		RuleCount:              len(cfg.Route.Rules),
-		RuleSetCount:           len(cfg.Route.RuleSet),
-		OutboundAWGCount:       awgCount,
-		OutboundCompositeCount: compCount,
-		Final:                  cfg.Route.Final,
-		FakeIPIface:            fakeIPIface,
-		FakeIPDns:              fakeIPDns,
-		FakeIPTunAddr:          fakeIPTunAddr,
-		Issues:                 issues,
-		LastError:              lastError,
-		CrashCount:             crashCount,
-		LastCrashReason:        lastCrashReason,
-		RestartSuppressedUntil: restartSuppressedUntil,
-	}, nil
-}
-
-func (s *ServiceImpl) Disable(ctx context.Context) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	// fakeip-tun teardown is an entirely separate path (no iptables; opkgtun +
-	// pool/CIDR routes + a fail-closed drain). Dispatch by mode before the
-	// tproxy body so the tproxy path below stays byte-for-byte unchanged for
-	// RoutingMode=="tproxy".
-	dispatchSettings, err := s.deps.Settings.Load()
-	if err != nil {
-		return err
-	}
-	// Dispatch on the RAW persisted RoutingMode, NOT the normalized value: a
-	// Normalize error (corrupt/hand-edited settings) would otherwise mis-route a
-	// fakeip-tun teardown into the tproxy body, orphaning the opkgtun/routes.
-	// A raw string compare keeps the fakeip branch independent of normalize.
-	if dispatchSettings.SingboxRouter.RoutingMode == "fakeip-tun" {
-		return s.disableFakeIPTun(ctx, dispatchSettings)
-	}
-
-	if err := s.deps.IPTables.Uninstall(ctx); err != nil {
-		s.appLog.Warn("uninstall", "", err.Error())
-	}
-	s.currentMark = ""
-	s.currentWANIPs = nil
-	s.currentLANBridges = nil
-	s.currentBypassPresets = nil
-	s.currentBypassExtraPorts = ""
-	s.currentBypassExtraSubnets = ""
-	s.currentIngress = nil
-	// Reset selective tracking too: after Uninstall the guard rules are gone,
-	// so the "currently applied" selective state is false. Leaving this stale
-	// made tproxy→off→tproxy re-enables skip the rebuild (selectiveChanged
-	// stayed false) and run with a permanently empty set.
-	s.currentSelectiveBypass = false
-	s.currentQoSClasses = nil
-	s.netfilterStateKnown = false
-	// Uninstall already tore down the fail-closed blackhole (if any); clear the
-	// tracking flag so a later reconcile doesn't try to remove it again.
-	s.blackholeActive = false
-
-	if s.deps.Orch != nil {
-		// Move 20-router.json under disabled/ — sing-box's non-recursive
-		// -C config.d does not see it after the next reload, so the
-		// tproxy inbound, route rules, DNS rules and composite outbounds
-		// all disappear from the merged config in one atomic rename.
-		if err := s.deps.Orch.SetEnabled(orchestrator.SlotRouter, false); err != nil {
-			s.appLog.Warn("orch-disable", "", err.Error())
-		}
-		// Park the QoS routes overlay with it: its rules reference qos-*
-		// inbound tags that only exist while 20-router.json is active.
-		if err := s.disableQoSRoutesSlot(); err != nil {
-			s.appLog.Warn("orch-disable", "qos-routes", err.Error())
-		}
-		// Композиты из 20-router.json только что пропали из merged-конфига.
-		// Синхронно даём device-proxy перегенерировать слот 30 (деградация
-		// до default-члена композита) — SetEnabled выше взвёл 250ms debounce,
-		// и один коалесцированный reload увидит уже корректный файл вместо
-		// того, чтобы prune молча вырезал vpn/vpn2 из селекторов (issue #465).
-		s.notifyRoutingSlotsChanged()
-	} else {
-		// Legacy fallback: strip the tproxy inbound in place so
-		// the running sing-box stops accepting on the TPROXY port
-		// after the persistConfigDirect reload.
-		cfg, err := s.loadRouterConfig()
-		if err == nil && cfg != nil {
-			filtered := make([]Inbound, 0, len(cfg.Inbounds))
-			for _, in := range cfg.Inbounds {
-				if in.Tag != "tproxy-in" {
-					filtered = append(filtered, in)
-				}
-			}
-			cfg.Inbounds = filtered
-			_ = s.persistConfigDirect(ctx, cfg)
-		}
-	}
-
-	settings, err := s.deps.Settings.Load()
-	if err != nil {
-		return err
-	}
-	settings.SingboxRouter.Enabled = false
-	if err := s.deps.Settings.Save(settings); err != nil {
-		return err
-	}
-
-	s.emitStatus(ctx)
-	return nil
-}
-
-func (s *ServiceImpl) Reconcile(ctx context.Context) error {
-	// A routing-mode switch (SwitchRoutingMode) holds transitionMu across its
-	// Disable→persist→Enable sequence, during which the persisted state is
-	// transiently half-flipped. Reconcile is a periodic heal — if a switch is in
-	// flight, skip this tick rather than act on the in-between state and race the
-	// switch's own (possibly rolling-back) Enable/Disable. TryLock: never block the
-	// scheduler; just defer the heal one tick.
-	if !s.transitionMu.TryLock() {
-		return nil
-	}
-	defer s.transitionMu.Unlock()
-	settings, err := s.deps.Settings.Load()
-	if err != nil {
-		return err
-	}
-	sr, err := NormalizeSingboxRouterSettings(settings.SingboxRouter)
-	if err != nil {
-		return err
-	}
-	// fakeip-tun installs NO iptables, so the tproxy switch below (keyed on
-	// IPTables.IsInstalled/HasAnyInstalled) would always read "not installed"
-	// and route every tick to Enable. Dispatch by mode FIRST so the tproxy
-	// switch stays byte-for-byte unchanged for RoutingMode=="tproxy".
-	if sr.RoutingMode == "fakeip-tun" {
-		return s.reconcileFakeIPTun(ctx, sr)
-	}
-	installedComplete := s.deps.IPTables.IsInstalled(ctx)
-	installedAny := s.deps.IPTables.HasAnyInstalled(ctx)
-	switch {
-	case sr.Enabled && !installedComplete:
-		// Drift-heal, NOT user-initiated: must honour a prior master-Stop, so
-		// do not clear the sticky intent (clearManualStop=false).
-		return s.enableLocked(ctx, false)
-	case !sr.Enabled && installedAny:
-		return s.Disable(ctx)
-	case sr.Enabled && installedComplete:
-		return s.reconcileInstalled(ctx, sr)
-	}
-	return nil
-}
-
-// reconcileInstalled handles the "Enabled && installed" branch:
-// detect mark or WAN-IP changes and re-Install. Extracted from Reconcile
-// to keep the decision tree testable without stubbing IsInstalled.
-func (s *ServiceImpl) reconcileInstalled(ctx context.Context, sr storage.SingboxRouterSettings) error {
-	sr, err := NormalizeSingboxRouterSettings(sr)
-	if err != nil {
-		return err
-	}
-	// Единственный рестарт-авторитет sing-box — watchdog (Operator.Reconcile,
-	// свой независимый 30s-тик). Router-reconcile больше НЕ рестартит движок
-	// сам: раньше это был второй независимый авторитет (#456), и вся
-	// токен-машинерия backoff'а существовала только чтобы примирить гонку двух
-	// рестартёров. Здесь лишь фиксируем факт смерти движка: engineDown ниже
-	// (а) включит fail-closed blackhole вместо перехвата в мёртвый порт,
-	// (б) погасит heal PREROUTING-джампов. Движок поднимет watchdog своим тиком;
-	// следующий reconcile-тик при живом движке восстановит перехват и снимет
-	// blackhole. Fail-closed держится blackhole'ом всё время, пока движок мёртв.
-	//
-	// «Готов» = process + оба inbound-сокета ПРИВЯЗАНЫ (singboxReady, hard-gate
-	// #221), а не просто IsRunning. Процесс может быть жив, но inbound не
-	// привязался (порт занят, отклонённый hot-reload) — тогда установка iptables
-	// REDIRECT/TPROXY в непривязанный сокет заблэкхолила бы весь policy-трафик,
-	// включая DNS:53. Поэтому up-but-unbound трактуем как engineDown: ставим
-	// fail-closed blackhole и НЕ ставим реальный перехват, пока сокеты не встанут.
-	engineReady := true
-	if s.deps.Singbox != nil {
-		engineReady = s.singboxReady(ctx, false)
-	}
-	engineDown := !engineReady
-	if sr.SelectiveBypass {
-		if err := s.validateSelectiveBypassAgainstApplied(sr); err != nil {
-			if errors.Is(err, errSelectiveIncompatible) {
-				// Definitive conflict with the APPLIED config — self-heal by
-				// persisting the disable so the guard doesn't blackhole traffic.
-				s.appLog.Warn("selective", "", err.Error()+"; disabling selective bypass")
-				settings, loadErr := s.deps.Settings.Load()
-				if loadErr == nil {
-					settings.SingboxRouter.SelectiveBypass = false
-					if saveErr := s.deps.Settings.Save(settings); saveErr != nil {
-						s.appLog.Warn("selective", "", "failed to persist selective bypass disable: "+saveErr.Error())
-					}
-				}
-				sr.SelectiveBypass = false
-			} else {
-				// Could not check (transient I/O, torn state during apply) —
-				// keep the user's setting and retry on the next tick. Flipping
-				// persisted settings on a read hiccup is not self-healing.
-				s.appLog.Warn("selective", "", err.Error()+"; keeping selective bypass, will re-validate")
-			}
-		}
-	}
-	policyMode := sr.DeviceMode == "" || sr.DeviceMode == "policy"
-	mark := ""
-	if policyMode {
-		mark, err = s.deps.Policies.GetPolicyMark(ctx, sr.PolicyName)
-		if err != nil || mark == "" {
-			// Policy gone upstream — fail-safe disable, no auto-recovery.
-			return s.Disable(ctx)
-		}
-	}
-	wanIPs, err := s.deps.WANIPCollector.Collect(ctx)
-	if err != nil {
-		return fmt.Errorf("collect WAN IPs: %w", err)
-	}
-
-	markChanged := mark != s.currentMark
-	wanIPsChanged := !slices.Equal(s.currentWANIPs, wanIPs)
-	var lanBridges []LANBridgeDNSRedir
-	if policyMode {
-		lanBridges, _ = DiscoverLANBridges(ctx, mark)
-	}
-	lanBridgesChanged := !equalLANBridges(s.currentLANBridges, lanBridges)
-	var ingress []string
-	if policyMode {
-		ingress = s.resolveIngressInterfaces(ctx, sr.IngressInterfaces)
-	}
-	ingressChanged := !slices.Equal(s.currentIngress, ingress)
-	bypassPresetsChanged := !slices.Equal(s.currentBypassPresets, sr.BypassPresets)
-	bypassExtraChanged := s.currentBypassExtraPorts != sr.BypassExtraPorts
-	bypassSubnetsChanged := s.currentBypassExtraSubnets != sr.BypassExtraSubnets
-
-	// Selective-bypass change detection.
-	selectiveChanged := s.currentSelectiveBypass != sr.SelectiveBypass
-
-	// QoS-DSCP change detection: only the iptables-relevant projection
-	// (DSCP + ports). An outbound-only change does not need an iptables
-	// re-Install — the healQoSConfig call below converges the sing-box side.
-	// Graceful degradation happens HERE, before change detection: when
-	// xt_dscp support is missing the desired dispatch set degrades to empty
-	// (feature-off; xtDscpUsable logs availability transitions) — gating
-	// later would leave desired≠installed forever and re-Install on every
-	// tick. When the module/extension shows up, the (TTL-bounded) probe
-	// flips and qosChanged triggers one re-Install with rules.
-	qosClasses := activeQoSClasses(sr.QoSClasses)
-	qosSpecs := qosIPTablesSpecs(qosClasses)
-	if len(qosSpecs) > 0 {
-		if err := EnsureXtDscpModule(ctx); err != nil {
-			s.appLog.Warn("ensure-xt-dscp", "", err.Error())
-		}
-		if !s.xtDscpUsable(ctx) {
-			qosSpecs = nil
-		}
-	}
-	qosChanged := !slices.Equal(s.currentQoSClasses, qosSpecs)
-
-	// Self-heal the sing-box side BEFORE any iptables change — same safe
-	// order as Enable (config → wait → Install). Installing new per-class
-	// dispatch ports first would blackhole class traffic onto ports nothing
-	// listens on until the debounced reload lands.
-	//
-	// healTProxyInbound: a previous Install rollback or upgrade hop may have
-	// left 20-router.json without the tproxy-in inbound — re-add it
-	// idempotently so sing-box keeps listening on TPROXYPort.
-	if err := s.healTProxyInbound(ctx, sr.UDPTimeout); err != nil {
-		s.appLog.Warn("heal-tproxy", "", err.Error())
-	}
-	// healQoSConfig: per-class inbound pairs (20-router.json) + managed route
-	// rules (18-qos-routes.json). Converges class add/remove/disable and
-	// outbound edits applied through UpdateSettings→Reconcile, and cleans
-	// stale qos-* artifacts. No-op (no write, no reload) when converged.
-	qosHealed := false
-	if healed, err := s.healQoSConfig(ctx, sr); err != nil {
-		s.appLog.Warn("heal-qos", "", err.Error())
-	} else {
-		qosHealed = healed
-	}
-	// The heal rewrote the QoS sing-box config AND the iptables port set is
-	// about to change: wait for sing-box to come back up on its inbounds
-	// before Install redirects traffic to the new per-class ports. Soft
-	// deadline — on timeout we proceed and accept the brief race rather
-	// than blocking the reconcile loop behind a dead engine forever.
-	if qosHealed && qosChanged {
-		if err := s.waitForSingbox(ctx, qosReloadWait); err != nil {
-			s.appLog.Warn("qos-dscp", "", fmt.Sprintf("sing-box not ready after QoS config heal: %v — installing anyway", err))
-		}
-	}
-
-	// After a daemon restart or upgrade the old awg-manager process died
-	// with no chance to run Uninstall, so stale AWGM chains, ip rules
-	// and ip routes may remain from the old process. netfilterStateKnown
-	// starts false on every fresh ServiceImpl, so the very first
-	// reconcileInstalled after startup always forces a full re-install
-	// regardless of what IsInstalled reports.
-	forceInitialSync := !s.netfilterStateKnown
-	// Self-heal: chains can survive while PREROUTING jumps get wiped (NDMS
-	// rebuilds PREROUTING on reconfig), leaving the engine "installed" but
-	// intercepting nothing. The netfilter.d hook restores them immediately on
-	// the NDMS reload; this is the slower secondary net. On a probe error treat
-	// the state as unknown and DO NOT reinstall — a transient `-S` failure
-	// during an NDMS reload must not trigger a needless rebuild.
-	_, jumps, probeErr := s.deps.IPTables.Probe(ctx)
-	jumpsMissing := probeErr == nil && !jumps
-	// wantBlackhole: движок мёртв И PREROUTING-джампы снесены (NDMS перестроил
-	// firewall). Раньше здесь перехват просто не восстанавливался в мёртвый порт
-	// (FIX-B), НО при снесённых джампах это означало fail-OPEN: policy-трафик
-	// уходил в обычный роутинг Keenetic → в WAN мимо proxy/AWG. Теперь ставим
-	// явный fail-closed blackhole — DROP policy-трафика (с теми же RETURN-
-	// исключениями LAN/router/WAN, что и у перехвата), чтобы гарантированно НЕ
-	// течь в WAN, пока движок не вернётся. Снимается ниже, когда движок оживёт.
-	wantBlackhole := jumpsMissing && engineDown
-	if wantBlackhole {
-		bypassUDP, bypassTCP, _ := resolveBypassPorts(sr.BypassPresets, sr.BypassExtraPorts)
-		bypassSubnets, _ := resolveBypassCIDRs(sr.BypassPresets, sr.BypassExtraSubnets)
-		// Selective guard references the AWGM-SELECTIVE ipset; ensure it exists so
-		// iptables-restore of the blackhole doesn't fail with "Set ... doesn't
-		// exist" (same pre-create the real Install path does below).
-		if sr.SelectiveBypass {
-			if e := ensureSelectiveSetExists(ctx); e != nil {
-				s.appLog.Warn("selective", "", fmt.Sprintf("pre-create ipset for blackhole: %v", e))
-			}
-		}
-		// Mirror the real interception spec's exclusions (bypass ports + selective
-		// guard) so the blackhole drops EXACTLY what would have been proxied — not
-		// the user's non-selective traffic and not their bypass ports.
-		blackholeSpec := RestoreInputSpec{
-			PolicyMark:     mark,
-			MatchAll:       !policyMode,
-			WANIPs:         wanIPs,
-			BypassCIDRs:    bypassSubnets,
-			BypassUDPPorts: bypassUDP,
-			BypassTCPPorts: bypassTCP,
-			SelectiveIPSet: sr.SelectiveBypass,
-		}
-		s.mu.Lock()
-		err := s.deps.IPTables.InstallBlackhole(ctx, blackholeSpec)
-		if err == nil {
-			s.blackholeActive = true
-		}
-		s.mu.Unlock()
-		if err != nil {
-			s.appLog.Warn("reconcile", "", "не удалось поставить fail-closed blackhole: "+err.Error())
-		} else {
-			s.appLog.Warn("reconcile", "", "движок не работает, PREROUTING jumps снесены — включён fail-closed blackhole (policy-трафик дропается, не течёт в WAN)")
-		}
-		// Реальный перехват в мёртвый порт всё равно не восстанавливаем.
-		jumpsMissing = false
-	}
-	needsInstall := forceInitialSync || jumpsMissing || markChanged || wanIPsChanged || lanBridgesChanged || ingressChanged || bypassPresetsChanged || bypassExtraChanged || bypassSubnetsChanged || selectiveChanged || qosChanged
-
-	// Движок не готов интерсептить (мёртв или inbound-сокеты не привязаны) —
-	// НЕ ставим iptables ни по какому триггеру (#221): REDIRECT/TPROXY в
-	// непривязанный сокет заблэкхолил бы весь policy-трафик, включая DNS:53.
-	// Fail-closed уже держит blackhole (при снесённых джампах) или перехват в
-	// мёртвый порт (при целых). Установку откладываем до готовности — следующий
-	// reconcile-тик поставит iptables, когда сокеты встанут.
-	if needsInstall && engineDown {
-		s.appLog.Warn("reconcile", "", "движок не готов (inbound-сокеты не привязаны) — откладываем установку iptables до готовности")
-		needsInstall = false
-	}
-
-	if needsInstall {
-		if forceInitialSync {
-			s.appLog.Info("reconcile", "", "first after daemon start — reinstalling netfilter rules")
-		} else if jumpsMissing {
-			s.appLog.Warn("reconcile", "", "PREROUTING jumps missing while chains present — reinstalling to restore interception")
-		}
-
-		if err := s.prepareNetfilter(ctx); err != nil {
-			return err
-		}
-
-		// If selective-bypass is enabled, the ipset MUST exist before
-		// iptables-restore runs — iptables-restore fails with "Set
-		// AWGM-SELECTIVE doesn't exist" if the set was never created.
-		// Create it empty now; it will be populated by the async rebuild
-		// triggered below. An empty set means no traffic is selectively
-		// intercepted yet, which is safe: the engine will fill it shortly.
-		if sr.SelectiveBypass {
-			if err := ensureSelectiveSetExists(ctx); err != nil {
-				s.appLog.Warn("selective", "", fmt.Sprintf("pre-create ipset failed: %v", err))
-				// Don't abort — if xt_set is missing, iptables-restore will
-				// surface a clear error; if ipset isn't installed but SelectiveBypass
-				// was somehow set, same. Proceed and let Install fail gracefully.
-			}
-		}
-
-		bypassUDP, bypassTCP, _ := resolveBypassPorts(sr.BypassPresets, sr.BypassExtraPorts)
-		bypassSubnets, _ := resolveBypassCIDRs(sr.BypassPresets, sr.BypassExtraSubnets)
-		s.mu.Lock()
-		if err := s.deps.IPTables.Install(ctx, RestoreInputSpec{
-			PolicyMark:        mark,
-			MatchAll:          !policyMode,
-			WANIPs:            wanIPs,
-			LANBridges:        lanBridges,
-			BypassUDPPorts:    bypassUDP,
-			BypassTCPPorts:    bypassTCP,
-			BypassCIDRs:       bypassSubnets,
-			IngressInterfaces: ingress,
-			SelectiveIPSet:    sr.SelectiveBypass,
-			QoSClasses:        qosSpecs,
-		}); err != nil {
-			s.mu.Unlock()
-			return err
-		}
-		s.currentMark = mark
-		s.currentWANIPs = wanIPs
-		s.currentLANBridges = lanBridges
-		s.currentBypassPresets = sr.BypassPresets
-		s.currentBypassExtraPorts = sr.BypassExtraPorts
-		s.currentBypassExtraSubnets = sr.BypassExtraSubnets
-		s.currentIngress = ingress
-		s.currentSelectiveBypass = sr.SelectiveBypass
-		s.currentQoSClasses = qosSpecs
-		s.netfilterStateKnown = true
-		s.mu.Unlock()
-
-		// If selective mode is being disabled, destroy the ipset so it
-		// doesn't linger in kernel memory.
-		if !sr.SelectiveBypass {
-			if err := destroySelectiveSet(ctx); err != nil {
-				s.appLog.Warn("selective", "", fmt.Sprintf("destroy ipset after disable: %v", err))
-			}
-			if err := s.disableSelectiveRoutesSlot(); err != nil {
-				s.appLog.Warn("selective", "", fmt.Sprintf("disable selective routes slot: %v", err))
-			}
-			if _, err := s.stripLegacySelectiveRulesFromRouter(ctx); err != nil {
-				s.appLog.Warn("selective", "", fmt.Sprintf("strip legacy selective rules: %v", err))
-			}
-		}
-	}
-
-	// Снимаем fail-closed blackhole ТОЛЬКО когда движок жив И probe успешен —
-	// тогда реальный перехват гарантированно на месте (steady state с целыми
-	// jumps, либо только что переустановлен выше; при ошибке Install был ранний
-	// return). Делаем ПОСЛЕ реального Install, чтобы не было окна утечки между
-	// снятием blackhole и восстановлением перехвата. Критично: на probe-ОШИБКЕ
-	// (jumpsMissing→false из-за !nil err) или мёртвом движке blackhole СОХРАНЯЕМ,
-	// иначе транзиентная -S ошибка во время NDMS-reload снесла бы DROP при живой
-	// утечке и удалила бы rules-файл, обездвижив и netfilter.d-хук. Идемпотентно.
-	if !engineDown && probeErr == nil {
-		s.mu.Lock()
-		if s.blackholeActive {
-			s.deps.IPTables.RemoveBlackhole(ctx)
-			s.blackholeActive = false
-			s.mu.Unlock()
-			s.appLog.Info("reconcile", "", "движок восстановлен — fail-closed blackhole снят")
-		} else {
-			s.mu.Unlock()
-		}
-	}
-
-	// Selective-bypass: rebuild on first Enable, when toggled on, or on daemon
-	// start only if the ipset was never populated yet. Rule changes trigger
-	// rebuild explicitly via staging Apply (frontend) — no periodic refresh.
-	if sr.SelectiveBypass && s.deps.SelectiveBuilder != nil {
-		configDir := ""
-		if s.deps.Singbox != nil {
-			configDir = s.deps.Singbox.ConfigDir()
-		}
-		needsInitialBuild := selective.NeedsPopulation(ctx, configDir)
-		if selectiveChanged || (forceInitialSync && needsInitialBuild) {
-			s.triggerSelectiveRebuild(ctx)
-		}
-	}
-	return nil
-}
-
 func (s *ServiceImpl) withConfig(ctx context.Context, event string, fn func(*RouterConfig) error) error {
 	cfg, err := s.loadRouterConfig()
 	if err != nil {
@@ -2201,1020 +783,6 @@ func (s *ServiceImpl) emitCfgEvent(event string, cfg *RouterConfig) {
 	}
 }
 
-// IsOutboundTagInUse reports whether tag is already occupied by any outbound
-// catalog visible to singbox-router.
-func (s *ServiceImpl) IsOutboundTagInUse(ctx context.Context, tag string) bool {
-	cfg, err := s.loadRouterConfig()
-	if err != nil {
-		cfg = NewEmptyConfig()
-	}
-	return s.isKnownOutboundTag(ctx, tag, cfg)
-}
-
-// RenameExternalOutboundTag rewrites references to a base outbound owned by
-// another producer (for example a single sing-box tunnel in 10-tunnels.json).
-// It updates both live/disabled router config and the pending draft, if any.
-func (s *ServiceImpl) RenameExternalOutboundTag(ctx context.Context, oldTag, newTag string) error {
-	oldTag = strings.TrimSpace(oldTag)
-	newTag = strings.TrimSpace(newTag)
-	if oldTag == "" || newTag == "" || oldTag == newTag {
-		return nil
-	}
-	// Settings-side references: QoS classes route to outbound tags too and
-	// must follow the rename (config-side rewrites below don't see them).
-	if err := s.renameQoSClassOutbound(oldTag, newTag); err != nil {
-		return err
-	}
-	if s.deps.Orch == nil {
-		return s.withConfig(ctx, "all", func(c *RouterConfig) error {
-			c.renameOutboundReferences(oldTag, newTag)
-			return nil
-		})
-	}
-
-	configDir := s.deps.Orch.ConfigDir()
-	activePath := filepath.Join(configDir, "20-router.json")
-	disabledPath := filepath.Join(configDir, "disabled", "20-router.json")
-	pendingPath := filepath.Join(configDir, "pending", "20-router.json")
-	changed := false
-
-	if data, ok, err := rewriteRouterConfigOutboundRefs(activePath, oldTag, newTag); err != nil {
-		return err
-	} else if ok {
-		if err := s.deps.Orch.Save(orchestrator.SlotRouter, data); err != nil {
-			return err
-		}
-		changed = true
-	}
-	if data, ok, err := rewriteRouterConfigOutboundRefs(disabledPath, oldTag, newTag); err != nil {
-		return err
-	} else if ok {
-		if err := storage.AtomicWrite(disabledPath, data); err != nil {
-			return err
-		}
-		changed = true
-	}
-	if data, ok, err := rewriteRouterConfigOutboundRefs(pendingPath, oldTag, newTag); err != nil {
-		return err
-	} else if ok {
-		if err := s.deps.Orch.SaveDraft(orchestrator.SlotRouter, data); err != nil {
-			return err
-		}
-		s.emitStagingEvent("staged")
-		changed = true
-	}
-	if changed {
-		if cfg, err := s.loadRouterConfig(); err == nil {
-			s.emitCfgEvent("all", s.ruleSetMaterializer().restoreConfig(cfg))
-		}
-	}
-	return nil
-}
-
-func rewriteRouterConfigOutboundRefs(path, oldTag, newTag string) ([]byte, bool, error) {
-	raw, err := os.ReadFile(path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil, false, nil
-		}
-		return nil, false, err
-	}
-	cfg := NewEmptyConfig()
-	if err := json.Unmarshal(raw, cfg); err != nil {
-		return nil, false, fmt.Errorf("parse %s: %w", path, err)
-	}
-	if len(cfg.outboundReferences(oldTag)) == 0 {
-		return nil, false, nil
-	}
-	cfg.renameOutboundReferences(oldTag, newTag)
-	data, err := json.MarshalIndent(cfg, "", "  ")
-	if err != nil {
-		return nil, false, err
-	}
-	return data, true, nil
-}
-
-func (s *ServiceImpl) ListRules(ctx context.Context) ([]Rule, error) {
-	cfg, err := s.loadRouterConfig()
-	if err != nil {
-		return nil, err
-	}
-	rules := s.ruleSetMaterializer().restoreConfig(cfg).Route.Rules
-	filtered, _ := stripSelectiveManagedRules(rules)
-	return filtered, nil
-}
-
-func (s *ServiceImpl) AddRule(ctx context.Context, r Rule) error {
-	return s.withConfig(ctx, "rules", func(c *RouterConfig) error { return c.AddRule(r) })
-}
-
-func (s *ServiceImpl) UpdateRule(ctx context.Context, index int, r Rule) error {
-	return s.withConfig(ctx, "rules", func(c *RouterConfig) error { return c.UpdateRule(index, r) })
-}
-
-func (s *ServiceImpl) DeleteRule(ctx context.Context, index int) error {
-	return s.withConfig(ctx, "rules", func(c *RouterConfig) error { return c.DeleteRule(index) })
-}
-
-func (s *ServiceImpl) MoveRule(ctx context.Context, from, to int) error {
-	return s.withConfig(ctx, "rules", func(c *RouterConfig) error { return c.MoveRule(from, to) })
-}
-
-func (s *ServiceImpl) SetRouteFinal(ctx context.Context, tag string) error {
-	if err := s.withConfig(ctx, "route", func(c *RouterConfig) error {
-		if !s.isKnownOutboundTag(ctx, tag, c) {
-			return fmt.Errorf("unknown outbound tag %q for route.final", tag)
-		}
-		return c.SetRouteFinal(tag)
-	}); err != nil {
-		return err
-	}
-	if tag != "direct" {
-		return s.disableSelectiveBypassIfEnabled(ctx)
-	}
-	return nil
-}
-
-// isKnownOutboundTag returns true if tag is a sing-box built-in or matches
-// an outbound from any known catalog (router composites, subscription
-// composites, AWG, sing-box tunnels).
-func (s *ServiceImpl) isKnownOutboundTag(ctx context.Context, tag string, cfg *RouterConfig) bool {
-	if tag == "direct" || tag == "block" || tag == "dns" {
-		return true
-	}
-	// Router-managed composites
-	for _, o := range cfg.Outbounds {
-		if o.Tag == tag {
-			return true
-		}
-	}
-	// Subscription composites (40-subscriptions.json)
-	if s.deps.SubscriptionComposites != nil {
-		for _, o := range s.deps.SubscriptionComposites.ListSubscriptionComposites() {
-			if o.Tag == tag {
-				return true
-			}
-		}
-	}
-	// AWG-direct outbounds (managed + system)
-	if s.deps.AWGTags != nil {
-		if tags, err := s.deps.AWGTags.ListTags(ctx); err == nil {
-			for _, t := range tags {
-				if t.Tag == tag {
-					return true
-				}
-			}
-		}
-	}
-	// Sing-box tunnels (10-tunnels.json)
-	if s.deps.SingboxTunnels != nil {
-		if tags, err := s.deps.SingboxTunnels.ListTunnelTags(ctx); err == nil {
-			for _, t := range tags {
-				if t == tag {
-					return true
-				}
-			}
-		}
-	}
-	return false
-}
-
-func (s *ServiceImpl) ListRuleSets(ctx context.Context) ([]RuleSet, error) {
-	cfg, err := s.loadRouterConfig()
-	if err != nil {
-		return nil, err
-	}
-	restored := s.ruleSetMaterializer().restoreConfig(cfg)
-	return restored.Route.RuleSet, nil
-}
-
-func (s *ServiceImpl) AddRuleSet(ctx context.Context, rs RuleSet) error {
-	if rs.Type == "" {
-		rs.Type = "remote"
-	}
-	if rs.Format == "" && rs.Type != "inline" {
-		rs.Format = "binary"
-	}
-	if rs.UpdateInterval == "" && rs.Type == "remote" {
-		rs.UpdateInterval = "24h"
-	}
-	return s.withConfig(ctx, "rulesets", func(c *RouterConfig) error { return c.AddRuleSet(rs) })
-}
-
-func (s *ServiceImpl) UpdateRuleSet(ctx context.Context, tag string, rs RuleSet) error {
-	if rs.Type == "" {
-		rs.Type = "remote"
-	}
-	if rs.Format == "" && rs.Type != "inline" {
-		rs.Format = "binary"
-	}
-	if rs.UpdateInterval == "" && rs.Type == "remote" {
-		rs.UpdateInterval = "24h"
-	}
-	return s.withConfig(ctx, "rulesets", func(c *RouterConfig) error { return c.UpdateRuleSet(tag, rs) })
-}
-
-func (s *ServiceImpl) DeleteRuleSet(ctx context.Context, tag string, force bool) error {
-	inlineTag := tag
-	if base, ok := inlineTagFromSRSTag(tag); ok {
-		inlineTag = base
-	}
-	return s.withConfig(ctx, "rulesets", func(c *RouterConfig) error {
-		if err := c.DeleteRuleSet(inlineTag, force); err != nil {
-			return err
-		}
-		if s.deps.Orch == nil {
-			s.ruleSetMaterializer().removeInlineArtifacts(inlineTag)
-		}
-		return nil
-	})
-}
-
-func (s *ServiceImpl) ListCompositeOutbounds(ctx context.Context) ([]CompositeOutboundView, error) {
-	cfg, err := s.loadRouterConfig()
-	if err != nil {
-		return nil, err
-	}
-	own := cfg.CompositeOutbounds()
-	out := make([]CompositeOutboundView, 0, len(own))
-	for _, o := range own {
-		out = append(out, CompositeOutboundView{Outbound: o, Source: "router"})
-	}
-	if s.deps.SubscriptionComposites != nil {
-		for _, o := range s.deps.SubscriptionComposites.ListSubscriptionComposites() {
-			out = append(out, CompositeOutboundView{Outbound: o, Source: "subscription"})
-		}
-	}
-	return out, nil
-}
-
-func (s *ServiceImpl) AddCompositeOutbound(ctx context.Context, o Outbound) error {
-	if strings.EqualFold(o.Type, "direct") {
-		if err := s.validateBindInterface(ctx, o.BindInterface); err != nil {
-			return err
-		}
-	}
-	return s.withConfig(ctx, "outbounds", func(c *RouterConfig) error { return c.AddCompositeOutbound(o) })
-}
-
-func (s *ServiceImpl) UpdateCompositeOutbound(ctx context.Context, tag string, o Outbound) error {
-	if strings.EqualFold(o.Type, "direct") {
-		if err := s.validateBindInterface(ctx, o.BindInterface); err != nil {
-			return err
-		}
-	}
-	if err := s.withConfig(ctx, "outbounds", func(c *RouterConfig) error { return c.UpdateCompositeOutbound(tag, o) }); err != nil {
-		return err
-	}
-	// A rename rewrites config references (renameOutboundReferences inside
-	// UpdateCompositeOutbound); mirror it for the settings-side QoS classes.
-	if newTag := strings.TrimSpace(o.Tag); newTag != "" && newTag != tag {
-		if err := s.renameQoSClassOutbound(tag, newTag); err != nil {
-			s.appLog.Warn("qos-classes", tag, "rename outbound in QoS classes: "+err.Error())
-		}
-	}
-	return nil
-}
-
-func (s *ServiceImpl) DeleteCompositeOutbound(ctx context.Context, tag string, force bool) error {
-	// QoS classes reference outbounds from settings, invisible to the config's
-	// own reference scan — guard them the same way route rules are guarded.
-	if !force {
-		if refs := s.qosClassesReferencing(tag); len(refs) > 0 {
-			return fmt.Errorf("%w: %q referenced by %s", ErrOutboundReferenced, tag, strings.Join(refs, ", "))
-		}
-	}
-	if err := s.withConfig(ctx, "outbounds", func(c *RouterConfig) error { return c.DeleteCompositeOutbound(tag, force) }); err != nil {
-		return err
-	}
-	if force {
-		// Force-delete DISABLES (not deletes) referencing classes so the UI
-		// shows them off instead of silently losing user configuration.
-		if err := s.disableQoSClassesForOutbound(tag); err != nil {
-			s.appLog.Warn("qos-classes", tag, "disable QoS classes after outbound delete: "+err.Error())
-		}
-	}
-	return nil
-}
-
-func (s *ServiceImpl) ListPresets() ([]Preset, error) {
-	return listRouterPresets(s.deps.PresetCatalog)
-}
-
-func (s *ServiceImpl) ApplyPreset(ctx context.Context, presetID, outboundTag string) error {
-	p, err := findRouterPreset(s.deps.PresetCatalog, presetID)
-	if err != nil {
-		return err
-	}
-	return s.withConfig(ctx, "status", func(c *RouterConfig) error {
-		return ApplyPresetToConfig(c, p, outboundTag)
-	})
-}
-
-func (s *ServiceImpl) GetSettings(ctx context.Context) (storage.SingboxRouterSettings, error) {
-	settings, err := s.deps.Settings.Load()
-	if err != nil {
-		return storage.SingboxRouterSettings{}, err
-	}
-	return NormalizeSingboxRouterSettings(settings.SingboxRouter)
-}
-
-func (s *ServiceImpl) UpdateSettings(ctx context.Context, sr storage.SingboxRouterSettings) error {
-	normalized, err := NormalizeSingboxRouterSettings(sr)
-	if err != nil {
-		return err
-	}
-	if err := s.validateSelectiveBypassSettings(ctx, normalized); err != nil {
-		return err
-	}
-	// QoS classes must route to outbounds that actually exist — an unknown
-	// tag would either be skipped at emit time (class silently inert) or,
-	// unguarded, take the whole merged config down at sing-box load. Checked
-	// here (not in the pure Normalize) because outbound catalogs are deps.
-	if err := s.validateQoSClassOutbounds(ctx, normalized.QoSClasses); err != nil {
-		return err
-	}
-	settings, err := s.deps.Settings.Load()
-	if err != nil {
-		return err
-	}
-	// Port-slot stability: the UI contract carries no slot field, so incoming
-	// classes are re-associated with their persisted slots by DSCP before the
-	// save — otherwise every PUT would re-deal ports positionally and RST
-	// untouched classes' flows.
-	normalized.QoSClasses = reassociateQoSSlots(normalized.QoSClasses, settings.SingboxRouter.QoSClasses)
-	settings.SingboxRouter = normalized
-	if err := s.deps.Settings.Save(settings); err != nil {
-		return err
-	}
-	return s.Reconcile(ctx)
-}
-
-// validateQoSClassOutbounds rejects ENABLED QoS classes whose Outbound does
-// not resolve against any known catalog (router composites, subscription
-// composites, AWG tags, sing-box tunnel tags, built-ins). Disabled classes
-// are deliberately exempt: outbound force-delete keeps the class around in a
-// disabled state (see disableQoSClassesForOutbound), and a later settings
-// PUT carrying that class verbatim must not 400. Wraps ErrQoSClassesInvalid
-// so the API maps it to 400 QOS_CLASSES_INVALID.
-func (s *ServiceImpl) validateQoSClassOutbounds(ctx context.Context, classes []storage.SingboxQoSClass) error {
-	if len(classes) == 0 {
-		return nil
-	}
-	cfg, err := s.loadRouterConfig()
-	if err != nil {
-		cfg = NewEmptyConfig()
-	}
-	for i, c := range classes {
-		if !c.Enabled {
-			continue
-		}
-		if !s.isKnownOutboundTag(ctx, c.Outbound, cfg) {
-			return fmt.Errorf("%w: qosClasses[%d]: неизвестный outbound %q", ErrQoSClassesInvalid, i, c.Outbound)
-		}
-	}
-	return nil
-}
-
-// ValidateSingboxRouterSettings enforces the WAN-binding discriminator:
-//   - WANAutoDetect=true   && WANInterface==""    → OK
-//   - WANAutoDetect=false  && WANInterface!=""    → OK
-//   - WANAutoDetect=true   && WANInterface!=""    → error (contradictory)
-//   - WANAutoDetect=false  && WANInterface==""    → error (no target)
-//
-// This guards both the storage layer (UpdateSettings) and the apply
-// path (Enable → EnsureRouteWAN) so an invalid state cannot reach
-// sing-box config either through the API or through a hand-edited
-// settings.json.
-func NormalizeSingboxRouterSettings(sr storage.SingboxRouterSettings) (storage.SingboxRouterSettings, error) {
-	if sr.DeviceMode == "" {
-		sr.DeviceMode = "policy"
-	}
-	switch sr.DeviceMode {
-	case "policy", "all":
-	default:
-		return sr, fmt.Errorf("deviceMode must be %q or %q, got %q", "policy", "all", sr.DeviceMode)
-	}
-	if sr.RoutingMode == "" {
-		sr.RoutingMode = "tproxy"
-	}
-	if sr.RoutingMode != "tproxy" && sr.RoutingMode != "fakeip-tun" {
-		return sr, fmt.Errorf("invalid routingMode %q (want tproxy|fakeip-tun)", sr.RoutingMode)
-	}
-	if sr.WANAutoDetect && sr.WANInterface != "" {
-		return sr, fmt.Errorf("wanAutoDetect=true requires wanInterface to be empty (got %q)", sr.WANInterface)
-	}
-	if !sr.WANAutoDetect && sr.WANInterface == "" {
-		return sr, fmt.Errorf("wanAutoDetect=false requires wanInterface to be set to a kernel interface name")
-	}
-	for _, name := range sr.BypassPresets {
-		if _, ok := knownPresets[name]; !ok {
-			return sr, fmt.Errorf("unknown bypass preset %q", name)
-		}
-	}
-	if _, _, err := parseExtraPorts(sr.BypassExtraPorts); err != nil {
-		return sr, fmt.Errorf("bypassExtraPorts: %w", err)
-	}
-	if _, err := resolveBypassSubnets(sr.BypassExtraSubnets); err != nil {
-		return sr, fmt.Errorf("bypassExtraSubnets: %w", err)
-	}
-	if err := validateIngressRefs(sr.IngressInterfaces); err != nil {
-		return sr, err
-	}
-	if err := normalizeFakeIPSettings(&sr); err != nil {
-		return sr, err
-	}
-	if sr.UDPTimeout != "" {
-		if _, err := time.ParseDuration(sr.UDPTimeout); err != nil {
-			return sr, fmt.Errorf("udpTimeout: invalid duration %q: %w", sr.UDPTimeout, err)
-		}
-	}
-	if err := validateQoSClasses(sr.QoSClasses); err != nil {
-		return sr, err
-	}
-	sr.QoSClasses = normalizeQoSClasses(sr.QoSClasses)
-	return sr, nil
-}
-
-// fakeIPMTUMin / fakeIPMTUMax bound the user-editable tun MTU. 576 is the IPv4
-// minimum-reassembly floor; 9000 is jumbo-frame ceiling. Outside this range
-// sing-tun behaves unpredictably, so reject early.
-const (
-	fakeIPMTUMin = 576
-	fakeIPMTUMax = 9000
-)
-
-// normalizeFakeIPSettings defaults the user-editable fakeip engine fields from
-// DefaultFakeIPTunParams (single source of truth) when empty/zero, then
-// validates them. Per spec, an empty FakeIPPool6 is defaulted to the v6 pool
-// (so a fresh install gets dual-stack); v6 is disabled at a higher layer, not
-// by persisting "" here. Idempotent: re-running on a normalized struct is a
-// fixed point.
-func normalizeFakeIPSettings(sr *storage.SingboxRouterSettings) error {
-	def := DefaultFakeIPTunParams()
-	if sr.FakeIPStack == "" {
-		sr.FakeIPStack = "gvisor"
-	}
-	if sr.FakeIPStack != "gvisor" && sr.FakeIPStack != "system" {
-		return fmt.Errorf("fakeipStack must be %q or %q, got %q", "gvisor", "system", sr.FakeIPStack)
-	}
-	if sr.FakeIPPool4 == "" {
-		sr.FakeIPPool4 = def.Inet4Range
-	}
-	if sr.FakeIPPool6 == "" {
-		sr.FakeIPPool6 = def.Inet6Range
-	}
-	if sr.FakeIPMTU == 0 {
-		sr.FakeIPMTU = def.MTU
-	}
-	// v4 pool must parse as an IPv4 prefix.
-	if p, err := netip.ParsePrefix(sr.FakeIPPool4); err != nil {
-		return fmt.Errorf("fakeipPool4: invalid CIDR %q: %w", sr.FakeIPPool4, err)
-	} else if !p.Addr().Is4() {
-		return fmt.Errorf("fakeipPool4: %q is not IPv4", sr.FakeIPPool4)
-	}
-	// v6 pool: empty disables v6; otherwise it must parse as an IPv6 prefix.
-	if sr.FakeIPPool6 != "" {
-		if p, err := netip.ParsePrefix(sr.FakeIPPool6); err != nil {
-			return fmt.Errorf("fakeipPool6: invalid CIDR %q: %w", sr.FakeIPPool6, err)
-		} else if p.Addr().Is4() {
-			return fmt.Errorf("fakeipPool6: %q is not IPv6", sr.FakeIPPool6)
-		}
-	}
-	if sr.FakeIPMTU < fakeIPMTUMin || sr.FakeIPMTU > fakeIPMTUMax {
-		return fmt.Errorf("fakeipMtu %d out of range [%d, %d]", sr.FakeIPMTU, fakeIPMTUMin, fakeIPMTUMax)
-	}
-	if sr.FakeIPRealServer == "" {
-		sr.FakeIPRealServer = def.RealServer
-	}
-	// Real upstream must be a plain IP: the fakeip topology resolves every
-	// domain through the "real" server itself, so a domain upstream could
-	// never bootstrap. Zoned addresses (fe80::1%eth0) make no sense for a
-	// DNS upstream either.
-	if addr, err := netip.ParseAddr(sr.FakeIPRealServer); err != nil {
-		return fmt.Errorf("fakeipRealServer: invalid IP address %q: %w", sr.FakeIPRealServer, err)
-	} else if addr.Zone() != "" {
-		return fmt.Errorf("fakeipRealServer: zoned address %q is not allowed", sr.FakeIPRealServer)
-	}
-	return nil
-}
-
-func validateIngressRefs(refs []string) error {
-	for _, ref := range refs {
-		if !strings.HasPrefix(ref, "managed:") && !strings.HasPrefix(ref, "iface:") {
-			return fmt.Errorf("ingress interface ref %q must be prefixed managed: or iface:", ref)
-		}
-		if strings.TrimSpace(strings.SplitN(ref, ":", 2)[1]) == "" {
-			return fmt.Errorf("ingress interface ref %q has empty target", ref)
-		}
-	}
-	return nil
-}
-
-func ValidateSingboxRouterSettings(sr storage.SingboxRouterSettings) error {
-	_, err := NormalizeSingboxRouterSettings(sr)
-	return err
-}
-
-func (s *ServiceImpl) ListWANInterfaces(ctx context.Context) ([]WANInterfaceInfo, error) {
-	if s.deps.WANInterfaces == nil {
-		return []WANInterfaceInfo{}, nil
-	}
-	return s.deps.WANInterfaces.ListWAN(ctx)
-}
-
-// ListBindableInterfaces returns interfaces a user can bind a direct
-// outbound to. Empty when no lister is wired.
-func (s *ServiceImpl) ListBindableInterfaces(ctx context.Context) ([]WANInterfaceInfo, error) {
-	if s.deps.BindableInterfaces == nil {
-		return []WANInterfaceInfo{}, nil
-	}
-	return s.deps.BindableInterfaces.ListBindable(ctx)
-}
-
-// ListIngressEligibleInterfaces возвращает интерфейсы, пригодные для
-// ingress-scope: bindable минус WAN минус LAN-бриджи (по Type). Для UI
-// router-страницы (мультиселект).
-func (s *ServiceImpl) ListIngressEligibleInterfaces(ctx context.Context) ([]WANInterfaceInfo, error) {
-	bindable, err := s.ListBindableInterfaces(ctx)
-	if err != nil {
-		return nil, err
-	}
-	wan, _ := s.ListWANInterfaces(ctx)
-	wanNames := map[string]bool{}
-	for _, w := range wan {
-		wanNames[w.Name] = true
-	}
-	out := make([]WANInterfaceInfo, 0, len(bindable))
-	for _, i := range bindable {
-		if wanNames[i.Name] || strings.EqualFold(i.Type, "Bridge") {
-			continue
-		}
-		out = append(out, i)
-	}
-	return out, nil
-}
-
-// validateBindInterface ensures name refers to a bindable interface. With
-// no lister wired (tests / minimal deployments) it is permissive.
-func (s *ServiceImpl) validateBindInterface(ctx context.Context, name string) error {
-	if s.deps.BindableInterfaces == nil {
-		return nil
-	}
-	ifaces, err := s.deps.BindableInterfaces.ListBindable(ctx)
-	if err != nil {
-		return err
-	}
-	for _, i := range ifaces {
-		if i.Name == name {
-			return nil
-		}
-	}
-	return fmt.Errorf("bind_interface %q is not a selectable interface", name)
-}
-
-func (s *ServiceImpl) computeIssues(cfg *RouterConfig) []Issue {
-	var issues []Issue
-	outboundTags := make(map[string]struct{})
-	for _, o := range cfg.Outbounds {
-		outboundTags[o.Tag] = struct{}{}
-	}
-	// AWG-direct outbounds live in 15-awg.json owned by awgoutbounds —
-	// add them to the validation set so rules referencing awg-{id} tags
-	// don't get flagged as orphans.
-	if s.deps.AWGTags != nil {
-		if awgTags, err := s.deps.AWGTags.ListTags(context.Background()); err == nil {
-			for _, t := range awgTags {
-				outboundTags[t.Tag] = struct{}{}
-			}
-		}
-	}
-	// Sing-box tunnels live in 10-tunnels.json owned by internal/singbox.
-	// Their tags (e.g. "veesp" for a VLESS outbound) are valid route
-	// targets but invisible to a router-only view of cfg.Outbounds.
-	if s.deps.SingboxTunnels != nil {
-		if tags, err := s.deps.SingboxTunnels.ListTunnelTags(context.Background()); err == nil {
-			for _, tag := range tags {
-				outboundTags[tag] = struct{}{}
-			}
-		}
-	}
-	// Subscription composites live in 40-subscriptions.json owned by subscription slot.
-	// Their tags are valid route targets but invisible to a router-only view of cfg.Outbounds.
-	if s.deps.SubscriptionComposites != nil {
-		for _, o := range s.deps.SubscriptionComposites.ListSubscriptionComposites() {
-			if o.Tag != "" {
-				outboundTags[o.Tag] = struct{}{}
-			}
-		}
-	}
-	for i, r := range cfg.Route.Rules {
-		issues = append(issues, s.computeRuleOutboundIssues(r, i, outboundTags)...)
-	}
-	if cfg.Route.Final != "" && !isKnownOutboundRef(cfg.Route.Final, outboundTags) {
-		issues = append(issues, Issue{
-			Severity: "warning",
-			Kind:     "orphan-outbound",
-			Tag:      cfg.Route.Final,
-			Message:  fmt.Sprintf("route.final ссылается на несуществующий outbound %q", cfg.Route.Final),
-		})
-	}
-	for i, o := range cfg.Outbounds {
-		for _, member := range o.Outbounds {
-			if !isKnownOutboundRef(member, outboundTags) {
-				issues = append(issues, Issue{
-					Severity: "warning",
-					Kind:     "orphan-outbound",
-					Tag:      member,
-					Message:  fmt.Sprintf("outbound %q содержит несуществующий member %q", o.Tag, member),
-				})
-			}
-		}
-		if o.Default != "" && !isKnownOutboundRef(o.Default, outboundTags) {
-			issues = append(issues, Issue{
-				Severity:  "warning",
-				Kind:      "orphan-outbound",
-				RuleIndex: i,
-				Tag:       o.Default,
-				Message:   fmt.Sprintf("outbound %q использует несуществующий default %q", o.Tag, o.Default),
-			})
-		}
-	}
-	for _, srv := range cfg.DNS.Servers {
-		if srv.Detour != "" && !isKnownOutboundRef(srv.Detour, outboundTags) {
-			issues = append(issues, Issue{
-				Severity: "warning",
-				Kind:     "orphan-outbound",
-				Tag:      srv.Detour,
-				Message:  fmt.Sprintf("DNS server %q использует несуществующий detour %q", srv.Tag, srv.Detour),
-			})
-		}
-	}
-	for _, rs := range cfg.Route.RuleSet {
-		if rs.DownloadDetour != "" && !isKnownOutboundRef(rs.DownloadDetour, outboundTags) {
-			issues = append(issues, Issue{
-				Severity: "warning",
-				Kind:     "orphan-outbound",
-				Tag:      rs.DownloadDetour,
-				Message:  fmt.Sprintf("rule_set %q использует несуществующий download_detour %q", rs.Tag, rs.DownloadDetour),
-			})
-		}
-	}
-
-	ruleSetTags := make(map[string]struct{}, len(cfg.Route.RuleSet))
-	for _, rs := range cfg.Route.RuleSet {
-		ruleSetTags[rs.Tag] = struct{}{}
-	}
-	for i, r := range cfg.Route.Rules {
-		issues = append(issues, computeRuleSetIssuesInRouteRule(r, i, ruleSetTags)...)
-	}
-	for i, r := range cfg.DNS.Rules {
-		for _, tag := range r.RuleSet {
-			if _, ok := ruleSetTags[tag]; !ok {
-				issues = append(issues, Issue{
-					Severity:  "warning",
-					Kind:      "orphan-rule-set",
-					RuleIndex: i,
-					Tag:       tag,
-					Message:   fmt.Sprintf("DNS-правило ссылается на несуществующий rule_set %q", tag),
-				})
-			}
-		}
-	}
-	return issues
-}
-
-func (s *ServiceImpl) computeRuleOutboundIssues(r Rule, index int, outboundTags map[string]struct{}) []Issue {
-	var issues []Issue
-	if r.Action == "route" && r.Outbound != "" && !isKnownOutboundRef(r.Outbound, outboundTags) {
-		issues = append(issues, Issue{
-			Severity:  "warning",
-			Kind:      "orphan-rule",
-			RuleIndex: index,
-			Tag:       r.Outbound,
-			Message:   fmt.Sprintf("правило ссылается на несуществующий outbound %q", r.Outbound),
-		})
-	}
-	for _, nested := range r.Rules {
-		issues = append(issues, s.computeRuleOutboundIssues(nested, index, outboundTags)...)
-	}
-	return issues
-}
-
-func isKnownOutboundRef(tag string, outboundTags map[string]struct{}) bool {
-	if tag == "direct" || tag == "block" || tag == "dns" {
-		return true
-	}
-	_, ok := outboundTags[tag]
-	return ok
-}
-
-func computeRuleSetIssuesInRouteRule(r Rule, index int, ruleSetTags map[string]struct{}) []Issue {
-	var issues []Issue
-	for _, tag := range r.RuleSet {
-		if _, ok := ruleSetTags[tag]; !ok {
-			issues = append(issues, Issue{
-				Severity:  "warning",
-				Kind:      "orphan-rule-set",
-				RuleIndex: index,
-				Tag:       tag,
-				Message:   fmt.Sprintf("правило ссылается на несуществующий rule_set %q", tag),
-			})
-		}
-	}
-	for _, nested := range r.Rules {
-		issues = append(issues, computeRuleSetIssuesInRouteRule(nested, index, ruleSetTags)...)
-	}
-	return issues
-}
-
-// RulesReferencing returns the indices of route rules whose outbound
-// equals tag. Used by tunnel.Service.Delete to refuse deletions that
-// would orphan references in router rules.
-func (s *ServiceImpl) RulesReferencing(tag string) []int {
-	cfg, err := s.loadRouterConfig()
-	if err != nil || cfg == nil {
-		return nil
-	}
-	return cfg.rulesReferencingOutbound(tag)
-}
-
-// OutboundReferenceLocations returns human-readable locations in the
-// router config that reference tag, EXCLUDING route.rules[...] (covered
-// by RulesReferencing). Used by the tunnel-delete guard to refuse
-// deletion of a tunnel still referenced via composite member, route
-// final, dns detour, or rule_set download_detour.
-func (s *ServiceImpl) OutboundReferenceLocations(tag string) []string {
-	cfg, err := s.loadRouterConfig()
-	if err != nil || cfg == nil {
-		return nil
-	}
-	return cfg.outboundReferencesExcludingRules(tag)
-}
-
-func (s *ServiceImpl) ListPolicies(ctx context.Context) ([]PolicyInfo, error) {
-	if s.deps.Policies == nil {
-		return nil, fmt.Errorf("access policy provider not configured")
-	}
-	return s.deps.Policies.ListPolicies(ctx)
-}
-
-func (s *ServiceImpl) CreatePolicy(ctx context.Context, description string) (PolicyInfo, error) {
-	if s.deps.Policies == nil {
-		return PolicyInfo{}, fmt.Errorf("access policy provider not configured")
-	}
-	if description == "" {
-		description = "awgm-router"
-	}
-	return s.deps.Policies.CreatePolicy(ctx, description)
-}
-
-func (s *ServiceImpl) ListPolicyDevices(ctx context.Context, policyName string) ([]PolicyDevice, error) {
-	if s.deps.Policies == nil {
-		return nil, fmt.Errorf("access policy provider not configured")
-	}
-	if policyName == "" {
-		return nil, fmt.Errorf("policy name required")
-	}
-	return s.deps.Policies.ListDevicesForPolicy(ctx, policyName)
-}
-
-func (s *ServiceImpl) BindDevice(ctx context.Context, mac, policyName string) error {
-	if s.deps.Policies == nil {
-		return fmt.Errorf("access policy provider not configured")
-	}
-	if mac == "" || policyName == "" {
-		return fmt.Errorf("mac and policyName required")
-	}
-	return s.deps.Policies.AssignDevice(ctx, mac, policyName)
-}
-
-func (s *ServiceImpl) UnbindDevice(ctx context.Context, mac string) error {
-	if s.deps.Policies == nil {
-		return fmt.Errorf("access policy provider not configured")
-	}
-	if mac == "" {
-		return fmt.Errorf("mac required")
-	}
-	return s.deps.Policies.UnassignDevice(ctx, mac)
-}
-
-// inspectSlotConfig returns the EFFECTIVE config the inspector must walk —
-// the slot that is live under the CURRENT routing mode — plus that slot for
-// draft reporting. Issue #488: the inspector always walked the tproxy slot
-// (20-router.json), so in fakeip-tun mode it explained decisions by DNS/route
-// rules and rule-set names sing-box wasn't even running; the live rules were
-// in the fakeip slot (21-fakeip.json).
-func (s *ServiceImpl) inspectSlotConfig() (*RouterConfig, orchestrator.Slot, error) {
-	mode := ""
-	if s.deps.Settings != nil {
-		settings, err := s.deps.Settings.Load()
-		if err != nil {
-			// Fail loud, not wrong: silently falling back to the tproxy slot
-			// on a transient settings read error would reproduce the very
-			// #488 bug (inspector explains decisions by the parked slot).
-			return nil, orchestrator.SlotRouter, fmt.Errorf("inspector: load settings: %w", err)
-		}
-		if settings != nil {
-			mode = settings.SingboxRouter.RoutingMode
-		}
-	}
-	slot := orchestrator.SlotRouter
-	if mode == "fakeip-tun" {
-		slot = orchestrator.SlotFakeIP
-	}
-	cfg, err := s.loadRouterConfigForMode(mode)
-	if err != nil {
-		return nil, slot, err
-	}
-	if cfg == nil {
-		cfg = NewEmptyConfig()
-	}
-	return cfg, slot, nil
-}
-
-// Inspect simulates which router rule would match the given input
-// (a domain or an IP). The matcher walk is purely Go; only rule_set
-// matchers shell out to `sing-box rule-set match` to consult the
-// binary or downloaded JSON list. Reads the persisted config of the
-// ACTIVE routing mode (tproxy or fakeip-tun slot) so the result reflects
-// what the user would observe at runtime.
-//
-// When the sing-box binary is unavailable (dev machine, fresh install
-// before the user has installed the package) rule_set matchers degrade
-// to no-match and a Note is appended to the result — the rest of the
-// inspector still works.
-func (s *ServiceImpl) Inspect(ctx context.Context, input InspectInput) (InspectResult, error) {
-	cfg, _, err := s.inspectSlotConfig()
-	if err != nil {
-		return InspectResult{}, err
-	}
-	final := cfg.Route.Final
-	if final == "" {
-		final = "direct"
-	}
-	binary := ""
-	if s.deps.Singbox != nil {
-		binary = s.deps.Singbox.Binary()
-	}
-	s.inspectCacheOnce.Do(func() {
-		s.inspectCache = newRuleSetCache("")
-	})
-	return Inspect(input, cfg.Route.Rules, cfg.Route.RuleSet, final, binary, s.inspectCache), nil
-}
-
-// InspectDNS simulates which DNS rule would match the given domain and how
-// the resolved DNS server classifies the resolution (fakeip → tunnel /
-// real → upstream / local → router). It is the DNS-resolution branch that
-// precedes the route inspector: a domain that gets a fakeip is then routed
-// by Inspect. The matcher walk is purely Go; only rule_set matchers shell
-// out to `sing-box rule-set match`. Reads the persisted config of the
-// ACTIVE routing mode (tproxy or fakeip-tun slot).
-func (s *ServiceImpl) InspectDNS(ctx context.Context, input InspectDNSInput) (InspectDNSResult, error) {
-	cfg, _, err := s.inspectSlotConfig()
-	if err != nil {
-		return InspectDNSResult{}, err
-	}
-	dnsRules := s.ruleSetMaterializer().restoreConfig(cfg).DNS.Rules
-	binary := ""
-	if s.deps.Singbox != nil {
-		binary = s.deps.Singbox.Binary()
-	}
-	s.inspectCacheOnce.Do(func() {
-		s.inspectCache = newRuleSetCache("")
-	})
-	return InspectDNS(input, dnsRules, cfg.DNS.Servers, cfg.Route.RuleSet, cfg.DNS.Final, binary, s.inspectCache), nil
-}
-
-func (s *ServiceImpl) InspectStream(ctx context.Context, input InspectInput) (<-chan InspectStreamEvent, error) {
-	ch := make(chan InspectStreamEvent, 32)
-	go func() {
-		defer close(ch)
-		emitEvent := func(ev InspectStreamEvent) bool {
-			select {
-			case <-ctx.Done():
-				return false
-			case ch <- ev:
-				return true
-			}
-		}
-		if !emitEvent(InspectStreamEvent{Type: "progress", Progress: &InspectProgress{Phase: "start", Message: "Запускаем инспектор маршрутов…"}}) {
-			return
-		}
-		if !emitEvent(InspectStreamEvent{Type: "progress", Progress: &InspectProgress{Phase: "load_config", Message: "Загружаем конфигурацию маршрутизации…"}}) {
-			return
-		}
-		cfg, slot, err := s.inspectSlotConfig()
-		if err != nil {
-			emitEvent(InspectStreamEvent{Type: "inspect-error", Error: err.Error()})
-			return
-		}
-		final := cfg.Route.Final
-		if final == "" {
-			final = "direct"
-		}
-		usingDraft := false
-		if s.deps.Orch != nil {
-			usingDraft = s.deps.Orch.DraftInfo(slot).HasDraft
-		}
-		if !emitEvent(InspectStreamEvent{Type: "progress", Progress: &InspectProgress{
-			Phase:        "config_loaded",
-			Message:      fmt.Sprintf("Конфигурация загружена: %d правил, %d rule_set, final: %s", len(cfg.Route.Rules), len(cfg.Route.RuleSet), final),
-			RuleTotal:    intPtr(len(cfg.Route.Rules)),
-			RuleSetTotal: intPtr(len(cfg.Route.RuleSet)),
-			Final:        final,
-			UsingDraft:   usingDraft,
-		}}) {
-			return
-		}
-		binary := ""
-		if s.deps.Singbox != nil {
-			binary = s.deps.Singbox.Binary()
-		}
-		s.inspectCacheOnce.Do(func() {
-			s.inspectCache = newRuleSetCache("")
-		})
-		res := InspectWithProgress(input, cfg.Route.Rules, cfg.Route.RuleSet, final, binary, s.inspectCache, func(p InspectProgress) {
-			select {
-			case <-ctx.Done():
-				return
-			case ch <- InspectStreamEvent{Type: "progress", Progress: &p}:
-			}
-		})
-		select {
-		case <-ctx.Done():
-			return
-		case ch <- InspectStreamEvent{Type: "result", Result: &res}:
-		}
-	}()
-	return ch, nil
-}
-
 // ---------------------------------------------------------------------------
 // Staging API
 // ---------------------------------------------------------------------------
-
-// StagingStatus is what /api/singbox/router/staging returns.
-type StagingStatus struct {
-	HasDraft   bool
-	DraftedAt  time.Time
-	Validation *orchestrator.ValidationResult
-}
-
-// StagingStatus returns metadata about the current pending draft for the
-// router slot. When a draft exists, Validation is populated with the
-// current cross-slot diagnostic so the UI can render a preview of "what
-// Apply would say".
-func (s *ServiceImpl) StagingStatus(ctx context.Context) StagingStatus {
-	info := s.deps.Orch.DraftInfo(orchestrator.SlotRouter)
-	st := StagingStatus{HasDraft: info.HasDraft, DraftedAt: info.DraftedAt}
-	if !info.HasDraft {
-		return st
-	}
-	bytes, err := s.deps.Orch.LoadEffective(orchestrator.SlotRouter)
-	if err != nil || bytes == nil {
-		return st
-	}
-	res := s.deps.Orch.ValidateDraft(orchestrator.SlotRouter, bytes)
-	st.Validation = &res
-	return st
-}
-
-// ApplyStaging is the service-level wrapper around Orch.ApplyDraft. On
-// success it emits "singbox.router.staging" + "singbox.router.rules" SSE
-// invalidations.
-func (s *ServiceImpl) ApplyStaging(ctx context.Context) (orchestrator.ValidationResult, error) {
-	_ = s.healLegacySelectiveRoutesSlotIfNeeded(ctx)
-	res, err := s.deps.Orch.ApplyDraft(orchestrator.SlotRouter)
-	if err == nil && res.Ok() {
-		// A staged rule-set delete/rename is final now — reap the orphaned
-		// inline/dat artifacts (issue #448: files were never deleted).
-		s.GCRuleSetArtifacts()
-		s.emitStagingEvent("applied")
-		s.emitRulesEvent()
-	}
-	return res, err
-}
-
-// DiscardStaging removes the pending draft for the router slot.
-func (s *ServiceImpl) DiscardStaging(ctx context.Context) error {
-	if err := s.deps.Orch.DiscardDraft(orchestrator.SlotRouter); err != nil {
-		return err
-	}
-	if err := s.restoreEffectiveRuleSetArtifacts(); err != nil {
-		return err
-	}
-	s.emitStagingEvent("discarded")
-	s.emitRulesEvent()
-	return nil
-}
-
-func (s *ServiceImpl) restoreEffectiveRuleSetArtifacts() error {
-	cfg, err := s.loadRouterConfig()
-	if err != nil {
-		return err
-	}
-	m := s.ruleSetMaterializer()
-	_, err = m.materializeConfig(m.restoreConfig(cfg))
-	return err
-}
